\name{ DFHCC3 }
\alias{ DFHCC3 }
\docType{data}
\title{ DFHCC3 }
\description{ ExpressionSet for the  DFHCC3  Dataset}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/16473279 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3744 
  PMIDs: 16473279 
  No abstract available.
  notes:
   summary:     
      Basal like cancerse_often lack an inactivated X chromosome.e_Other markers
 found were duplication of the active X chromosome ande_nonheterochromatin
ized X chromosomal DNA. A small subset of X chromosomal genes were overexp
ressed. These abnormalities are thought to led to the pathogenesis of basa
l like cancers.
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID
   preprocessing:     
      As published by original author.

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 1007_s_at 1053_at ... AFFX-HUMISGF3A/M97935_MB_at
    (42447 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 42447 features, 40 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
   Length     Class      Mode 
       40 character character 

alt_sample_name: 
   Length     Class      Mode 
       40 character character 

sample_type: 
tumor 
   40 

batch: 
DFHCC3 
    40 

uncurated_author_metadata: 
   Length     Class      Mode 
       40 character character 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3744 }
\keyword{datasets}
