\name{ DFHCC3_eset }
\alias{ DFHCC3_eset }
\docType{data}
\title{  }
\usage{data( DFHCC3_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/16473279 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3744 
  PMIDs: 16473279 
  No abstract available.
  notes:
   summary:[1] "DUKE"
[1] "DUKE2"
[1] "DUKE2"
[1] "EMC2"
[1] "EORTC10994"
[1] "EXPO"
[1] "FNCLCC"
[1] "HLP"
[1] "IRB"
[1] "KOO"
[1] "LUND"
[1] "LUND2"
[1] "MAINZ"
[1] "MAQC2"
[1] "MCCC"
[1] "MDA4"
[1] "METABRIC"
[1] "MSK"
[1] "MUG"
[1] "NCCS"
[1] "NCI"
[1] "NKI"
[1] "PNC"
[1] "STK"
[1] "STNO2"
[1] "SUPERTAM_HGU133A"
[1] "SUPERTAM_HGU133PLUS2"
[1] "TCGA"
[1] "TRANSBIG"
[1] "UCSF"
[1] "UNC4"
[1] "UNT"
[1] "UPP"
[1] "VDX"
