\name{ GSE32646 }
\alias{ GSE32646 }
\docType{data}
\title{ GSE32646 }
\description{ ExpressionSet for the  GSE32646  Dataset}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information:  
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE32646 
  PMIDs: 22320227 
  No abstract available.
  notes:
   summary:     
      
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID
   preprocessing:     
      As published by original author.

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 1007_s_at 1053_at ... 91952_at (42437 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 42437 features, 115 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
   Length     Class      Mode 
      115 character character 

sample_type: 
tumor 
  115 

er: 
negative positive 
      44       71 

pgr: 
negative positive 
      70       45 

her2: 
negative positive 
      81       34 

T: 
 1  2  3  4 
 5 87 18  5 

N: 
 0  1 
32 83 

age_at_initial_pathologic_diagnosis: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  27.00   45.00   51.00   51.49   59.00   73.00 

grade: 
 1  2  3 
16 78 21 

batch: 
GSE32646 
     115 

uncurated_author_metadata: 
   Length     Class      Mode 
      115 character character 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE32646 }
\keyword{datasets}
