\name{ GSE58644 }
\alias{ GSE58644 }
\docType{data}
\title{ GSE58644 }
\description{ ExpressionSet for the  GSE58644  Dataset}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information:  
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE58644 
  PMIDs: 25284793 
  No abstract available.
  notes:
   summary:     
      
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID
   preprocessing:     
      As published by original author.

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 7896756 7896759 ... 8180179 (21462 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 21462 features, 321 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
   Length     Class      Mode 
      321 character character 

alt_sample_name: 
   Length     Class      Mode 
      321 character character 

sample_type: 
tumor 
  321 

er: 
negative positive     NA's 
      70      250        1 

her2: 
negative positive     NA's 
     256       58        7 

tumor_size: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  0.600   1.600   2.100   2.354   2.600  15.000 

T: 
   1    2    3    4 NA's 
  43   59   13    1  205 

N: 
   0    1 NA's 
 138  151   32 

age_at_initial_pathologic_diagnosis: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  29.00   49.00   58.00   58.82   68.00   93.00 

grade: 
   1    2    3 NA's 
  26  135  159    1 

dmfs_status: 
norecurrence   recurrence 
         295           26 

dmfs_days: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
      0    9496   17900   21620   33600   52590 

treatment: 
chemo.plus.hormono       chemotherapy     hormonotherapy          untreated 
                91                 29                 66                 10 
              NA's 
               125 

chemo: 
   0    1 NA's 
 105  123   93 

tamoxifen: 
   0    1 NA's 
  39  157  125 

herceptin: 
   0    1 NA's 
 190   12  119 

batch: 
GSE58644 
     321 

uncurated_author_metadata: 
   Length     Class      Mode 
      321 character character 

duplicates: 
   Length     Class      Mode 
      321 character character 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE58644 }
\keyword{datasets}
