\name{ HLP_eset }
\alias{ HLP_eset }
\docType{data}
\title{  }
\usage{data( HLP_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/?term=19688261 
  Title:  
  URL: http://www.ebi.ac.uk/arrayexpress/experiments/E-TABM-543/ 
  PMIDs: 19688261 
  No abstract available.
  notes:
   summary:     
      The results show evidence of different patterns of genetic aberrations in distinct molecular subtypes of breast cancer. P
atterns of copy number aberrations may drive biological phenomena characteristic to each subtype.
   version:     
      2015-04-27 19:09:29
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 9g8cQB1TZtuiix.ulU fJUdX0IAn_P9VLTgJU ... xopB7pPn18FJ067uDs (26536 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 26536 features, 53 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
 HLP_B1 HLP_B11 HLP_B12 HLP_B15 HLP_B16 HLP_B17  HLP_B2 HLP_B20 HLP_B21 HLP_B24 HLP_B26 HLP_B27 HLP_B29 HLP_B30 HLP_B31 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
HLP_B32 HLP_B33 HLP_B34 HLP_B35 HLP_B37 HLP_B39 HLP_B40 HLP_B41 HLP_B43 HLP_B44 HLP_B45 HLP_B46 HLP_B47 HLP_B49  HLP_B5 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
 HLP_B6 HLP_B60 HLP_B61 HLP_B62 HLP_B64 HLP_B65 HLP_B66 HLP_B67 HLP_B69  HLP_B7 HLP_B72 HLP_B74 HLP_B76 HLP_B77 HLP_B78 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
HLP_B79 HLP_B80 HLP_B81 HLP_B82 HLP_B84 HLP_B86 HLP_B87  HLP_B9 
      1       1       1       1       1       1       1       1 

alt_sample_name: 
 B1 B11 B12 B15 B16 B17  B2 B20 B21 B24 B26 B27 B29 B30 B31 B32 B33 B34 B35 B37 B39 B40 B41 B43 B44 B45 B46 B47 B49  B5  B6 
  1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
B60 B61 B62 B64 B65 B66 B67 B69  B7 B72 B74 B76 B77 B78 B79 B80 B81 B82 B84 B86 B87  B9 
  1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 

sample_type: 
tumor 
   53 

er: 
negative positive 
      28       25 

pgr: 
negative positive 
      33       20 

her2: 
negative positive 
      40       13 

tumor_size: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
  1.200   1.800   2.450   2.648   3.000   8.000       5 

N: 
   0    1 NA's 
  27   25    1 

age_at_initial_pathologic_diagnosis: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
  30.00   47.50   53.50   54.96   64.25   81.00       5 

grade: 
 3 
53 

batch: 
HLP 
 53 

uncurated_author_metadata: 
  samplename: HLP_B1///dataset: HLP///series: HLP///id: B1///er: 1///pgr: 0///her2: 0///size: 1.5///node: 0///age: 51///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B11///dataset: HLP///series: HLP///id: B11///er: 0///pgr: 1///her2: 0///size: 1.7///node: 0///age: 57///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B12///dataset: HLP///series: HLP///id: B12///er: 0///pgr: 0///her2: 1///size: 3///node: 1///age: 53///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B15///dataset: HLP///series: HLP///id: B15///er: 1///pgr: 1///her2: 0///size: 1.7///node: 1///age: 38///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B16///dataset: HLP///series: HLP///id: B16///er: 1///pgr: 1///her2: 0///size: 2.5///node: 1///age: 57///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B17///dataset: HLP///series: HLP///id: B17///er: 0///pgr: 0///her2: 0///size: 2.3///node: 0///age: 53///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B2///dataset: HLP///series: HLP///id: B2///er: 0///pgr: 0///her2: 0///size: 2.5///node: 1///age: 48///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B20///dataset: HLP///series: HLP///id: B20///er: 0///pgr: 0///her2: 1///size: 1.8///node: 0///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B21///dataset: HLP///series: HLP///id: B21///er: 1///pgr: 1///her2: 0///size: 4///node: 1///age: 61///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B24///dataset: HLP///series: HLP///id: B24///er: 1///pgr: 0///her2: 0///size: 1.5///node: 0///age: 67///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B26///dataset: HLP///series: HLP///id: B26///er: 1///pgr: 0///her2: 1///size: 2///node: 0///age: 37///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B27///dataset: HLP///series: HLP///id: B27///er: 1///pgr: 1///her2: 0///size: 2.7///node: 1///age: 43///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
 samplename: HLP_B29///dataset: HLP///series: HLP///id: B29///er: 0///pgr: 0///her2: 0///size: NA///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B30///dataset: HLP///series: HLP///id: B30///er: 1///pgr: 1///her2: 0///size: 1.8///node: 0///age: 44///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B31///dataset: HLP///series: HLP///id: B31///er: 1///pgr: 1///her2: 0///size: 1.7///node: 0///age: 53///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B32///dataset: HLP///series: HLP///id: B32///er: 0///pgr: 0///her2: 1///size: 1.8///node: 0///age: 49///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B33///dataset: HLP///series: HLP///id: B33///er: 1///pgr: 1///her2: 0///size: 2///node: 1///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B34///dataset: HLP///series: HLP///id: B34///er: 0///pgr: 0///her2: 1///size: 5///node: 1///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
 samplename: HLP_B35///dataset: HLP///series: HLP///id: B35///er: 0///pgr: 0///her2: 1///size: 3///node: NA///age: 80///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B37///dataset: HLP///series: HLP///id: B37///er: 1///pgr: 1///her2: 0///size: 2.5///node: 1///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B39///dataset: HLP///series: HLP///id: B39///er: 1///pgr: 1///her2: 0///size: 3.5///node: 1///age: 54///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B40///dataset: HLP///series: HLP///id: B40///er: 0///pgr: 1///her2: 1///size: 3.7///node: 1///age: 72///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B41///dataset: HLP///series: HLP///id: B41///er: 0///pgr: 0///her2: 0///size: 1.7///node: 0///age: 70///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B43///dataset: HLP///series: HLP///id: B43///er: 0///pgr: 0///her2: 1///size: 2///node: 1///age: 59///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B44///dataset: HLP///series: HLP///id: B44///er: 1///pgr: 1///her2: 0///size: 2.5///node: 1///age: 43///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B45///dataset: HLP///series: HLP///id: B45///er: 0///pgr: 0///her2: 0///size: 3.8///node: 1///age: 52///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B46///dataset: HLP///series: HLP///id: B46///er: 0///pgr: 0///her2: 1///size: 2.2///node: 0///age: 49///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B47///dataset: HLP///series: HLP///id: B47///er: 1///pgr: 0///her2: 0///size: 1.7///node: 0///age: 65///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B49///dataset: HLP///series: HLP///id: B49///er: 1///pgr: 1///her2: 0///size: 2///node: 0///age: 59///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
    samplename: HLP_B5///dataset: HLP///series: HLP///id: B5///er: 1///pgr: 1///her2: 0///size: 8///node: 1///age: 81///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B6///dataset: HLP///series: HLP///id: B6///er: 0///pgr: 0///her2: 0///size: 3.5///node: 0///age: 69///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B60///dataset: HLP///series: HLP///id: B60///er: 1///pgr: 1///her2: 0///size: 1.5///node: 0///age: 63///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B61///dataset: HLP///series: HLP///id: B61///er: 1///pgr: 1///her2: 0///size: 2.7///node: 1///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B62///dataset: HLP///series: HLP///id: B62///er: 0///pgr: 0///her2: 0///size: 3.5///node: 1///age: 50///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B64///dataset: HLP///series: HLP///id: B64///er: 0///pgr: 0///her2: 0///size: 2.1///node: 0///age: 53///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B65///dataset: HLP///series: HLP///id: B65///er: 0///pgr: 0///her2: 0///size: 1.2///node: 0///age: 46///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B66///dataset: HLP///series: HLP///id: B66///er: 0///pgr: 0///her2: 1///size: 3///node: 0///age: 37///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
 samplename: HLP_B67///dataset: HLP///series: HLP///id: B67///er: 0///pgr: 0///her2: 0///size: NA///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
 samplename: HLP_B69///dataset: HLP///series: HLP///id: B69///er: 0///pgr: 0///her2: 0///size: NA///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B7///dataset: HLP///series: HLP///id: B7///er: 1///pgr: 0///her2: 0///size: 2.7///node: 1///age: 70///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B72///dataset: HLP///series: HLP///id: B72///er: 1///pgr: 1///her2: 0///size: 2.4///node: 0///age: 64///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B74///dataset: HLP///series: HLP///id: B74///er: 1///pgr: 0///her2: 0///size: 2.8///node: 1///age: 41///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B76///dataset: HLP///series: HLP///id: B76///er: 1///pgr: 1///her2: 1///size: 2.5///node: 0///age: 68///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B77///dataset: HLP///series: HLP///id: B77///er: 1///pgr: 1///her2: 1///size: 7///node: 1///age: 49///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B78///dataset: HLP///series: HLP///id: B78///er: 0///pgr: 0///her2: 0///size: 2.2///node: 0///age: 35///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B79///dataset: HLP///series: HLP///id: B79///er: 0///pgr: 0///her2: 0///size: 2.4///node: 1///age: 50///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B80///dataset: HLP///series: HLP///id: B80///er: 0///pgr: 0///her2: 1///size: 3///node: 0///age: 30///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B81///dataset: HLP///series: HLP///id: B81///er: 0///pgr: 0///her2: 0///size: 2///node: 0///age: 30///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B82///dataset: HLP///series: HLP///id: B82///er: 1///pgr: 1///her2: 0///size: 1.2///node: 0///age: 67///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
 samplename: HLP_B84///dataset: HLP///series: HLP///id: B84///er: 0///pgr: 0///her2: 0///size: NA///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
 samplename: HLP_B86///dataset: HLP///series: HLP///id: B86///er: 0///pgr: 0///her2: 0///size: NA///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
samplename: HLP_B87///dataset: HLP///series: HLP///id: B87///er: 0///pgr: 0///her2: 0///size: 2.7///node: 1///age: 61///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 
  samplename: HLP_B9///dataset: HLP///series: HLP///id: B9///er: 1///pgr: 0///her2: 0///size: 2.6///node: 1///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                   1 

}}
\source{ http://www.ebi.ac.uk/arrayexpress/experiments/E-TABM-543/ }
\keyword{datasets}
