\name{ KOO }
\alias{ KOO }
\docType{data}
\title{ KOO }
\description{ link does not work, in progress8 }
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/12747878 
  Title:  
  URL: Unavailable 
  PMIDs: 12747878 

  Abstract: A 6 word abstract is available. Use 'abstract' method.
  notes:
   summary:     
      A new gene signature was used to accurately predict 90% of relapse cases i
n the study.
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID
   preprocessing:     
      As published by original author.

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 1007_s_at 1053_at ... AFFX-HUMISGF3A/M97935_MB_at (280
    total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 280 features, 88 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
   Length     Class      Mode 
       88 character character 

alt_sample_name: 
   Length     Class      Mode 
       88 character character 

sample_type: 
tumor 
   88 

er: 
negative positive 
      15       73 

pgr: 
negative positive 
      23       65 

tumor_size: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
   0.20    1.80    2.30    2.74    3.50    8.50 

N: 
 0  1 
19 69 

treatment: 
chemotherapy    untreated 
          61           27 

batch: 
KOO 
 88 

uncurated_author_metadata: 
   Length     Class      Mode 
       88 character character 

duplicates: 
   Length     Class      Mode 
       88 character character 

}}
\source{ Unavailable }
\keyword{datasets}
