\name{ LUND2_eset }
\alias{ LUND2_eset }
\docType{data}
\title{  }
\usage{data( LUND2_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/?term=17452630 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE5325 
  PMIDs: 17452630 
  No abstract available.
  notes:
   summary:
featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 1 2 ... 27648 (22008 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 22008 features, 105 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
LUND2_Ca10109 LUND2_Ca10289 LUND2_Ca10335 LUND2_Ca10419 LUND2_Ca10426 LUND2_Ca10481 LUND2_Ca10580 LUND2_Ca10591 LUND2_Ca10607 
            1             1             1             1             1             1             1             1             1 
LUND2_Ca10879 LUND2_Ca10960 LUND2_Ca11205 LUND2_Ca11379 LUND2_Ca11442 LUND2_Ca11460 LUND2_Ca11522 LUND2_Ca11570 LUND2_Ca11591 
            1             1             1             1             1             1             1             1             1 
LUND2_Ca11620 LUND2_Ca11829 LUND2_Ca11968 LUND2_Ca13351  LUND2_Ca6012  LUND2_Ca6014  LUND2_Ca6021  LUND2_Ca6042  LUND2_Ca6061 
            1             1             1             1             1             1             1             1             1 
 LUND2_Ca6110  LUND2_Ca6125  LUND2_Ca6144  LUND2_Ca6146  LUND2_Ca6165  LUND2_Ca6188  LUND2_Ca6206  LUND2_Ca6219  LUND2_Ca6257 
            1             1             1             1             1             1             1             1             1 
 LUND2_Ca6318  LUND2_Ca6366  LUND2_Ca6388  LUND2_Ca6464  LUND2_Ca6474  LUND2_Ca6516  LUND2_Ca6537  LUND2_Ca6617  LUND2_Ca6687 
            1             1             1             1             1             1             1             1             1 
 LUND2_Ca6716  LUND2_Ca6773  LUND2_Ca7102  LUND2_Ca7106  LUND2_Ca7151  LUND2_Ca7168  LUND2_Ca7170  LUND2_Ca7214  LUND2_Ca7244 
            1             1             1             1             1             1             1             1             1 
 LUND2_Ca7246  LUND2_Ca7357  LUND2_Ca7366  LUND2_Ca7367  LUND2_Ca7398  LUND2_Ca7408  LUND2_Ca7422  LUND2_Ca7461  LUND2_Ca7493 
            1             1             1             1             1             1             1             1             1 
 LUND2_Ca7495  LUND2_Ca7549  LUND2_Ca7583  LUND2_Ca7589  LUND2_Ca7604  LUND2_Ca7609  LUND2_Ca7632  LUND2_Ca7722  LUND2_Ca7732 
            1             1             1             1             1             1             1             1             1 
 LUND2_Ca7811  LUND2_Ca7845  LUND2_Ca7921  LUND2_Ca7982  LUND2_Ca8173  LUND2_Ca8174  LUND2_Ca8196  LUND2_Ca8212  LUND2_Ca8237 
            1             1             1             1             1             1             1             1             1 
 LUND2_Ca8290  LUND2_Ca8333  LUND2_Ca8401  LUND2_Ca8477  LUND2_Ca8537  LUND2_Ca8560  LUND2_Ca8582  LUND2_Ca8610  LUND2_Ca8676 
            1             1             1             1             1             1             1             1             1 
 LUND2_Ca8800  LUND2_Ca8860  LUND2_Ca9077  LUND2_Ca9079  LUND2_Ca9126  LUND2_Ca9159  LUND2_Ca9174  LUND2_Ca9298  LUND2_Ca9539 
            1             1             1             1             1             1             1             1             1 
      (Other) 
            6 

alt_sample_name: 
Ca10109 Ca10289 Ca10335 Ca10419 Ca10426 Ca10481 Ca10580 Ca10591 Ca10607 Ca10879 Ca10960 Ca11205 Ca11379 Ca11442 Ca11460 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
Ca11522 Ca11570 Ca11591 Ca11620 Ca11829 Ca11968 Ca13351  Ca6012  Ca6014  Ca6021  Ca6042  Ca6061  Ca6110  Ca6125  Ca6144 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
 Ca6146  Ca6165  Ca6188  Ca6206  Ca6219  Ca6257  Ca6318  Ca6366  Ca6388  Ca6464  Ca6474  Ca6516  Ca6537  Ca6617  Ca6687 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
 Ca6716  Ca6773  Ca7102  Ca7106  Ca7151  Ca7168  Ca7170  Ca7214  Ca7244  Ca7246  Ca7357  Ca7366  Ca7367  Ca7398  Ca7408 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
 Ca7422  Ca7461  Ca7493  Ca7495  Ca7549  Ca7583  Ca7589  Ca7604  Ca7609  Ca7632  Ca7722  Ca7732  Ca7811  Ca7845  Ca7921 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
 Ca7982  Ca8173  Ca8174  Ca8196  Ca8212  Ca8237  Ca8290  Ca8333  Ca8401  Ca8477  Ca8537  Ca8560  Ca8582  Ca8610  Ca8676 
      1       1       1       1       1       1       1       1       1       1       1       1       1       1       1 
 Ca8800  Ca8860  Ca9077  Ca9079  Ca9126  Ca9159  Ca9174  Ca9298  Ca9539 (Other) 
      1       1       1       1       1       1       1       1       1       6 

sample_type: 
tumor 
  105 

er: 
negative positive 
      60       45 

treatment: 
hormonotherapy 
           105 

batch: 
LUND2 
  105 

uncurated_author_metadata: 
samplename: LUND2_Ca10109///dataset: LUND2///series: LUND2///id: Ca10109///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10289///dataset: LUND2///series: LUND2///id: Ca10289///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10335///dataset: LUND2///series: LUND2///id: Ca10335///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10419///dataset: LUND2///series: LUND2///id: Ca10419///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10426///dataset: LUND2///series: LUND2///id: Ca10426///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10481///dataset: LUND2///series: LUND2///id: Ca10481///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10580///dataset: LUND2///series: LUND2///id: Ca10580///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10591///dataset: LUND2///series: LUND2///id: Ca10591///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10607///dataset: LUND2///series: LUND2///id: Ca10607///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10879///dataset: LUND2///series: LUND2///id: Ca10879///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca10960///dataset: LUND2///series: LUND2///id: Ca10960///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11205///dataset: LUND2///series: LUND2///id: Ca11205///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11379///dataset: LUND2///series: LUND2///id: Ca11379///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11442///dataset: LUND2///series: LUND2///id: Ca11442///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11460///dataset: LUND2///series: LUND2///id: Ca11460///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11522///dataset: LUND2///series: LUND2///id: Ca11522///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11570///dataset: LUND2///series: LUND2///id: Ca11570///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11591///dataset: LUND2///series: LUND2///id: Ca11591///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11620///dataset: LUND2///series: LUND2///id: Ca11620///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11829///dataset: LUND2///series: LUND2///id: Ca11829///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca11968///dataset: LUND2///series: LUND2///id: Ca11968///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
samplename: LUND2_Ca13351///dataset: LUND2///series: LUND2///id: Ca13351///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6012///dataset: LUND2///series: LUND2///id: Ca6012///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6014///dataset: LUND2///series: LUND2///id: Ca6014///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6021///dataset: LUND2///series: LUND2///id: Ca6021///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6042///dataset: LUND2///series: LUND2///id: Ca6042///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6061///dataset: LUND2///series: LUND2///id: Ca6061///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6110///dataset: LUND2///series: LUND2///id: Ca6110///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6125///dataset: LUND2///series: LUND2///id: Ca6125///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6144///dataset: LUND2///series: LUND2///id: Ca6144///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6146///dataset: LUND2///series: LUND2///id: Ca6146///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6165///dataset: LUND2///series: LUND2///id: Ca6165///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6188///dataset: LUND2///series: LUND2///id: Ca6188///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6206///dataset: LUND2///series: LUND2///id: Ca6206///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6219///dataset: LUND2///series: LUND2///id: Ca6219///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6257///dataset: LUND2///series: LUND2///id: Ca6257///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6318///dataset: LUND2///series: LUND2///id: Ca6318///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6366///dataset: LUND2///series: LUND2///id: Ca6366///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6388///dataset: LUND2///series: LUND2///id: Ca6388///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6464///dataset: LUND2///series: LUND2///id: Ca6464///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6474///dataset: LUND2///series: LUND2///id: Ca6474///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6516///dataset: LUND2///series: LUND2///id: Ca6516///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6537///dataset: LUND2///series: LUND2///id: Ca6537///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6617///dataset: LUND2///series: LUND2///id: Ca6617///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6687///dataset: LUND2///series: LUND2///id: Ca6687///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6716///dataset: LUND2///series: LUND2///id: Ca6716///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca6773///dataset: LUND2///series: LUND2///id: Ca6773///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7102///dataset: LUND2///series: LUND2///id: Ca7102///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7106///dataset: LUND2///series: LUND2///id: Ca7106///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7151///dataset: LUND2///series: LUND2///id: Ca7151///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7168///dataset: LUND2///series: LUND2///id: Ca7168///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7170///dataset: LUND2///series: LUND2///id: Ca7170///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7214///dataset: LUND2///series: LUND2///id: Ca7214///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7244///dataset: LUND2///series: LUND2///id: Ca7244///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7246///dataset: LUND2///series: LUND2///id: Ca7246///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7357///dataset: LUND2///series: LUND2///id: Ca7357///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7366///dataset: LUND2///series: LUND2///id: Ca7366///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7367///dataset: LUND2///series: LUND2///id: Ca7367///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7398///dataset: LUND2///series: LUND2///id: Ca7398///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7408///dataset: LUND2///series: LUND2///id: Ca7408///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7422///dataset: LUND2///series: LUND2///id: Ca7422///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7461///dataset: LUND2///series: LUND2///id: Ca7461///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7493///dataset: LUND2///series: LUND2///id: Ca7493///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7495///dataset: LUND2///series: LUND2///id: Ca7495///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7549///dataset: LUND2///series: LUND2///id: Ca7549///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7583///dataset: LUND2///series: LUND2///id: Ca7583///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7589///dataset: LUND2///series: LUND2///id: Ca7589///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7604///dataset: LUND2///series: LUND2///id: Ca7604///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7609///dataset: LUND2///series: LUND2///id: Ca7609///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7632///dataset: LUND2///series: LUND2///id: Ca7632///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7722///dataset: LUND2///series: LUND2///id: Ca7722///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7732///dataset: LUND2///series: LUND2///id: Ca7732///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7811///dataset: LUND2///series: LUND2///id: Ca7811///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7845///dataset: LUND2///series: LUND2///id: Ca7845///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7921///dataset: LUND2///series: LUND2///id: Ca7921///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca7982///dataset: LUND2///series: LUND2///id: Ca7982///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8173///dataset: LUND2///series: LUND2///id: Ca8173///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8174///dataset: LUND2///series: LUND2///id: Ca8174///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8196///dataset: LUND2///series: LUND2///id: Ca8196///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8212///dataset: LUND2///series: LUND2///id: Ca8212///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8237///dataset: LUND2///series: LUND2///id: Ca8237///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8290///dataset: LUND2///series: LUND2///id: Ca8290///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8333///dataset: LUND2///series: LUND2///id: Ca8333///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8401///dataset: LUND2///series: LUND2///id: Ca8401///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8477///dataset: LUND2///series: LUND2///id: Ca8477///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8537///dataset: LUND2///series: LUND2///id: Ca8537///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8560///dataset: LUND2///series: LUND2///id: Ca8560///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8582///dataset: LUND2///series: LUND2///id: Ca8582///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8610///dataset: LUND2///series: LUND2///id: Ca8610///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8676///dataset: LUND2///series: LUND2///id: Ca8676///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8800///dataset: LUND2///series: LUND2///id: Ca8800///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca8860///dataset: LUND2///series: LUND2///id: Ca8860///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca9077///dataset: LUND2///series: LUND2///id: Ca9077///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca9079///dataset: LUND2///series: LUND2///id: Ca9079///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca9126///dataset: LUND2///series: LUND2///id: Ca9126///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca9159///dataset: LUND2///series: LUND2///id: Ca9159///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca9174///dataset: LUND2///series: LUND2///id: Ca9174///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca9298///dataset: LUND2///series: LUND2///id: Ca9298///er: 0///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
  samplename: LUND2_Ca9539///dataset: LUND2///series: LUND2///id: Ca9539///er: 1///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 2 
                                                                                                                                                                                                                                                   1 
                                                                                                                                                                                                                                             (Other) 
                                                                                                                                                                                                                                                   6 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE5325 }
\keyword{datasets}
