\name{ MCCC_eset }
\alias{ MCCC_eset }
\docType{data}
\title{  }
\usage{data( MCCC_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/?term=19960244 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE19177 
  PMIDs: 19960244 
  No abstract available.
  notes:
   summary:     
      Overall, expression and copy number profiling of familial tumors have shown that the tumors show molecular heterogeneity 
similar to sporadic tumors and are defined by their molecular subtypes rather than BRCA1 or BRCA2 germline mutation statu
s. 
   version:     
      2015-04-27 19:18:33
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: probe_10017 probe_10021 ... probe_7650767 (19048 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 19048 features, 75 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
MCCC_GSM475354 MCCC_GSM475355 MCCC_GSM475356 MCCC_GSM475357 MCCC_GSM475358 MCCC_GSM475359 MCCC_GSM475360 MCCC_GSM475361 
             1              1              1              1              1              1              1              1 
MCCC_GSM475362 MCCC_GSM475363 MCCC_GSM475364 MCCC_GSM475365 MCCC_GSM475366 MCCC_GSM475367 MCCC_GSM475368 MCCC_GSM475369 
             1              1              1              1              1              1              1              1 
MCCC_GSM475370 MCCC_GSM475371 MCCC_GSM475372 MCCC_GSM475373 MCCC_GSM475374 MCCC_GSM475375 MCCC_GSM475376 MCCC_GSM475377 
             1              1              1              1              1              1              1              1 
MCCC_GSM475378 MCCC_GSM475379 MCCC_GSM475380 MCCC_GSM475381 MCCC_GSM475382 MCCC_GSM475383 MCCC_GSM475384 MCCC_GSM475385 
             1              1              1              1              1              1              1              1 
MCCC_GSM475386 MCCC_GSM475387 MCCC_GSM475388 MCCC_GSM475389 MCCC_GSM475390 MCCC_GSM475391 MCCC_GSM475392 MCCC_GSM475393 
             1              1              1              1              1              1              1              1 
MCCC_GSM475394 MCCC_GSM475395 MCCC_GSM475396 MCCC_GSM475397 MCCC_GSM475398 MCCC_GSM475399 MCCC_GSM475400 MCCC_GSM475401 
             1              1              1              1              1              1              1              1 
MCCC_GSM475402 MCCC_GSM475403 MCCC_GSM475404 MCCC_GSM475405 MCCC_GSM475406 MCCC_GSM475407 MCCC_GSM475408 MCCC_GSM475409 
             1              1              1              1              1              1              1              1 
MCCC_GSM475410 MCCC_GSM475411 MCCC_GSM475412 MCCC_GSM475413 MCCC_GSM475414 MCCC_GSM475415 MCCC_GSM475416 MCCC_GSM475417 
             1              1              1              1              1              1              1              1 
MCCC_GSM475418 MCCC_GSM475419 MCCC_GSM475420 MCCC_GSM475421 MCCC_GSM475422 MCCC_GSM475423 MCCC_GSM475424 MCCC_GSM475425 
             1              1              1              1              1              1              1              1 
MCCC_GSM475426 MCCC_GSM475427 MCCC_GSM475428 
             1              1              1 

sample_type: 
tumor 
   75 

batch: 
MCCC 
  75 

uncurated_author_metadata: 
samplename: MCCC_GSM475354///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475355///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475356///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475357///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475358///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475359///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475360///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475361///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475362///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475363///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475364///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475365///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475366///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475367///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475368///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475369///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475370///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475371///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475372///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475373///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475374///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475375///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475376///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475377///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475378///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475379///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475380///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475381///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475382///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475383///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475384///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475385///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475386///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475387///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475388///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475389///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475390///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475391///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475392///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475393///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475394///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475395///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475396///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475397///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475398///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475399///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475400///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475401///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475402///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475403///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475404///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475405///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475406///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475407///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475408///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475409///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475410///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475411///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475412///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475413///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475414///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475415///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475416///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475417///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475418///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475419///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475420///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475421///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475422///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475423///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475424///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475425///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475426///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475427///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 
samplename: MCCC_GSM475428///dataset: MCCC///series: MCCC///id: NA///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                               1 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE19177 }
\keyword{datasets}
