\name{ MUG }
\alias{ MUG }
\docType{data}
\title{ MUG }
\description{ ExpressionSet for the  MUG  Dataset}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/?term=18592372 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=gse10510 
  PMIDs: 18592372 
  No abstract available.
  notes:
   summary:     
      A method was developed to separate tumor cells and their microenvironment 
to test the prognostic abilities of the immune system. Results showed that
 lymphatic infiltration is beneficial for ER negative patients, but probab
ly not beneficial for ER positive patients.
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID
   preprocessing:     
      As published by original author.

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: H200000001 H200000005 ... opHsV04TC000043 (14288 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 14288 features, 152 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
   Length     Class      Mode 
      152 character character 

alt_sample_name: 
   Length     Class      Mode 
      152 character character 

sample_type: 
tumor 
  152 

batch: 
MUG 
152 

uncurated_author_metadata: 
   Length     Class      Mode 
      152 character character 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=gse10510 }
\keyword{datasets}
