\name{ DFHCC3_eset }
\alias{ DFHCC3_eset }
\docType{data}
\title{ DFHCC3_eset }
\usage{data( DFHCC3_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/16473279 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3744 
  PMIDs: 16473279 
  No abstract available.
  notes:
   summary:
featureData([1] "DUKE"
bject of class 'AnnotatedDataFrame'
  featureNames: 1007_s_at 1053_at ... AFFX-HUMISGF3A/M97935_MB_at (42447 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 42447 features, 40 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
DFHCC3_T115 DFHCC3_T116 DFHCC3_T117 DFHCC3_T118 DFHCC3_T119 DFHCC3_T123 DFHCC3_T129 DFHCC3_T130 DFHCC3_T133 DFHCC3_T134 
          1           1           1           1           1           1           1           1           1           1 
DFHCC3_T137 DFHCC3_T140 DFHCC3_T141 DFHCC3_T143 DFHCC3_T144 DFHCC3_T145 DFHCC3_T146 DFHCC3_T147 DFHCC3_T149 DFHCC3_T151 
          1           1           1           1           1           1           1           1           1           1 
DFHCC3_T152 DFHCC3_T161 DFHCC3_T162 DFHCC3_T175 DFHCC3_T178 DFHCC3_T183  DFHCC3_T21  DFHCC3_T30  DFHCC3_T37  DFHCC3_T38 
          1           1           1           1           1           1           1           1           1           1 
  DFHCC3_T4  DFHCC3_T41  DFHCC3_T44  DFHCC3_T50  DFHCC3_T56  DFHCC3_T73  DFHCC3_T74  DFHCC3_T81  DFHCC3_T84  DFHCC3_T92 
          1           1           1           1           1           1           1           1           1           1 

alt_sample_name: 
T115 T116 T117 T118 T119 T123 T129 T130 T133 T134 T137 T140 T141 T143 T144 T145 T146 T147 T149 T151 T152 T161 T162 T175 T178 
   1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 
T183  T21  T30  T37  T38   T4  T41  T44  T50  T56  T73  T74  T81  T84  T92 
   1    1    1    1    1    1    1    1    1    1    1    1    1    1    1 

sample_type: 
tumor 
   40 

batch: 
DFHCC3 
    40 

uncurated_author_metadata: 
samplename: DFHCC3_T115///dataset: DFHCC3///series: DFHCC3///id: T115///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T116///dataset: DFHCC3///series: DFHCC3///id: T116///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T117///dataset: DFHCC3///series: DFHCC3///id: T117///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T118///dataset: DFHCC3///series: DFHCC3///id: T118///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T119///dataset: DFHCC3///series: DFHCC3///id: T119///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T123///dataset: DFHCC3///series: DFHCC3///id: T123///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T129///dataset: DFHCC3///series: DFHCC3///id: T129///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T130///dataset: DFHCC3///series: DFHCC3///id: T130///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T133///dataset: DFHCC3///series: DFHCC3///id: T133///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T134///dataset: DFHCC3///series: DFHCC3///id: T134///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T137///dataset: DFHCC3///series: DFHCC3///id: T137///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T140///dataset: DFHCC3///series: DFHCC3///id: T140///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T141///dataset: DFHCC3///series: DFHCC3///id: T141///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T143///dataset: DFHCC3///series: DFHCC3///id: T143///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T144///dataset: DFHCC3///series: DFHCC3///id: T144///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T145///dataset: DFHCC3///series: DFHCC3///id: T145///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T146///dataset: DFHCC3///series: DFHCC3///id: T146///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T147///dataset: DFHCC3///series: DFHCC3///id: T147///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T149///dataset: DFHCC3///series: DFHCC3///id: T149///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T151///dataset: DFHCC3///series: DFHCC3///id: T151///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T152///dataset: DFHCC3///series: DFHCC3///id: T152///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T161///dataset: DFHCC3///series: DFHCC3///id: T161///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T162///dataset: DFHCC3///series: DFHCC3///id: T162///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T175///dataset: DFHCC3///series: DFHCC3///id: T175///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T178///dataset: DFHCC3///series: DFHCC3///id: T178///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
samplename: DFHCC3_T183///dataset: DFHCC3///series: DFHCC3///id: T183///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T21///dataset: DFHCC3///series: DFHCC3///id: T21///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T30///dataset: DFHCC3///series: DFHCC3///id: T30///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T37///dataset: DFHCC3///series: DFHCC3///id: T37///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T38///dataset: DFHCC3///series: DFHCC3///id: T38///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
    samplename: DFHCC3_T4///dataset: DFHCC3///series: DFHCC3///id: T4///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T41///dataset: DFHCC3///series: DFHCC3///id: T41///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T44///dataset: DFHCC3///series: DFHCC3///id: T44///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T50///dataset: DFHCC3///series: DFHCC3///id: T50///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T56///dataset: DFHCC3///series: DFHCC3///id: T56///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T73///dataset: DFHCC3///series: DFHCC3///id: T73///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T74///dataset: DFHCC3///series: DFHCC3///id: T74///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T81///dataset: DFHCC3///series: DFHCC3///id: T81///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T84///dataset: DFHCC3///series: DFHCC3///id: T84///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 
  samplename: DFHCC3_T92///dataset: DFHCC3///series: DFHCC3///id: T92///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                                  1 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3744 }
\keyword{datasets}
