\name{ EXPO_eset }
\alias{ EXPO_eset }
\docType{data}
\title{ EXPO_eset }
\usage{data( EXPO_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information:  
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE2109 
  PMIDs:  
  No abstract available.
  notes:
   summary:     
      N/A
   version:     
      2015-04-27 19:07:04
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 1007_s_at 1053_at ... AFFX-HUMISGF3A/M97935_MB_at (42447 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 42447 features, 353 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
GSM102439 GSM102440 GSM102441 GSM102468 GSM102482 GSM102487 GSM102499 GSM102502 GSM102506 GSM102523 GSM102525 GSM102526 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM102530 GSM102534 GSM102536 GSM102544 GSM102545 GSM102563 GSM102564 GSM102566 GSM102567 GSM102569 GSM102571 GSM102573 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM102576 GSM102580 GSM102583 GSM117578 GSM117580 GSM117581 GSM117584 GSM117585 GSM117588 GSM117639 GSM117659 GSM117660 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM117661 GSM117675 GSM117685 GSM117686 GSM117687 GSM117692 GSM117695 GSM117719 GSM117730 GSM117745 GSM117758 GSM117759 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM117761 GSM117766 GSM117777 GSM137895 GSM137896 GSM137898 GSM137900 GSM137903 GSM137924 GSM137925 GSM137936 GSM137937 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM137939 GSM137943 GSM137944 GSM137950 GSM137953 GSM137963 GSM137978 GSM137984 GSM137987 GSM137988 GSM137989 GSM137991 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM137994 GSM138011 GSM138014 GSM138027 GSM138028 GSM138031 GSM138035 GSM138041 GSM138049 GSM138051 GSM138054 GSM152564 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM152569 GSM152570 GSM152571 GSM152576 GSM152585 GSM152590 GSM152605 GSM152615 GSM152616 GSM152618 GSM152620 GSM152621 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM152627 GSM152645 GSM152661   (Other) 
        1         1         1       254 

alt_sample_name: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
   1005   21640  101100  134700  215900  486200 

sample_type: 
tumor 
  353 

er: 
negative positive     NA's 
      85      161      107 

pgr: 
negative positive     NA's 
     114      129      110 

her2: 
negative positive     NA's 
     166       61      126 

age_at_initial_pathologic_diagnosis: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
  25.00   45.00   55.00   59.44   67.50   95.00       1 

grade: 
   1    2    3 NA's 
  32  114  151   56 

batch: 
EXPO 
 353 

uncurated_author_metadata: 
samplename: GSM102439///dataset: EXPO///series: EXPO///id: 101135///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102440///dataset: EXPO///series: EXPO///id: 101141///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
  samplename: GSM102441///dataset: EXPO///series: EXPO///id: 101145///er: 0///pgr: 0///her2: NA///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102468///dataset: EXPO///series: EXPO///id: 104438///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 35///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM102482///dataset: EXPO///series: EXPO///id: 118668///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM102487///dataset: EXPO///series: EXPO///id: 127822///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102499///dataset: EXPO///series: EXPO///id: 129692///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102502///dataset: EXPO///series: EXPO///id: 129702///er: 1///pgr: 0///her2: 1///size: NA///node: NA///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102506///dataset: EXPO///series: EXPO///id: 133638///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 35///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102523///dataset: EXPO///series: EXPO///id: 149054///er: 1///pgr: 0///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102525///dataset: EXPO///series: EXPO///id: 149057///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM102526///dataset: EXPO///series: EXPO///id: 151146///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 45///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102530///dataset: EXPO///series: EXPO///id: 152477///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 85///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102534///dataset: EXPO///series: EXPO///id: 161433///er: 0///pgr: 0///her2: 1///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102536///dataset: EXPO///series: EXPO///id: 161439///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM102544///dataset: EXPO///series: EXPO///id: 180960///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM102545///dataset: EXPO///series: EXPO///id: 182732///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
 samplename: GSM102563///dataset: EXPO///series: EXPO///id: 78407///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
     samplename: GSM102564///dataset: EXPO///series: EXPO///id: 8032///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM102566///dataset: EXPO///series: EXPO///id: 84136///er: 1///pgr: 0///her2: 0///size: NA///node: NA///age: 75///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
 samplename: GSM102567///dataset: EXPO///series: EXPO///id: 84138///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 35///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM102569///dataset: EXPO///series: EXPO///id: 84141///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM102571///dataset: EXPO///series: EXPO///id: 84144///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 45///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM102573///dataset: EXPO///series: EXPO///id: 84148///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 35///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM102576///dataset: EXPO///series: EXPO///id: 84154///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 65///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM102580///dataset: EXPO///series: EXPO///id: 84174///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM102583///dataset: EXPO///series: EXPO///id: 89267///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117578///dataset: EXPO///series: EXPO///id: 104351///er: 0///pgr: 0///her2: 1///size: NA///node: NA///age: 65///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117580///dataset: EXPO///series: EXPO///id: 112875///er: 1///pgr: 0///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117581///dataset: EXPO///series: EXPO///id: 112876///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117584///dataset: EXPO///series: EXPO///id: 129693///er: 1///pgr: 0///her2: 0///size: NA///node: NA///age: 65///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117585///dataset: EXPO///series: EXPO///id: 129701///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 35///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117588///dataset: EXPO///series: EXPO///id: 133647///er: 1///pgr: 0///her2: 1///size: NA///node: NA///age: 45///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM117639///dataset: EXPO///series: EXPO///id: 143654///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 75///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117659///dataset: EXPO///series: EXPO///id: 149051///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117660///dataset: EXPO///series: EXPO///id: 149056///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117661///dataset: EXPO///series: EXPO///id: 151602///er: 0///pgr: 0///her2: 1///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117675///dataset: EXPO///series: EXPO///id: 152481///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 85///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117685///dataset: EXPO///series: EXPO///id: 161430///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117686///dataset: EXPO///series: EXPO///id: 161435///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117687///dataset: EXPO///series: EXPO///id: 161462///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 45///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117692///dataset: EXPO///series: EXPO///id: 161516///er: 0///pgr: 0///her2: 1///size: NA///node: NA///age: 65///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117695///dataset: EXPO///series: EXPO///id: 161521///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
  samplename: GSM117719///dataset: EXPO///series: EXPO///id: 174937///er: 1///pgr: 1///her2: NA///size: NA///node: NA///age: 55///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117730///dataset: EXPO///series: EXPO///id: 180955///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117745///dataset: EXPO///series: EXPO///id: 184839///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117758///dataset: EXPO///series: EXPO///id: 187213///er: 0///pgr: 0///her2: 1///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117759///dataset: EXPO///series: EXPO///id: 187214///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 55///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM117761///dataset: EXPO///series: EXPO///id: 187217///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM117766///dataset: EXPO///series: EXPO///id: 21775///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM117777///dataset: EXPO///series: EXPO///id: 84189///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 55///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM137895///dataset: EXPO///series: EXPO///id: 112874///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 35///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137896///dataset: EXPO///series: EXPO///id: 112877///er: 1///pgr: 0///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137898///dataset: EXPO///series: EXPO///id: 127825///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 25///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
 samplename: GSM137900///dataset: EXPO///series: EXPO///id: 127837///er: 0///pgr: 0///her2: NA///size: NA///node: NA///age: 75///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137903///dataset: EXPO///series: EXPO///id: 133640///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 65///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137924///dataset: EXPO///series: EXPO///id: 177892///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 35///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137925///dataset: EXPO///series: EXPO///id: 177893///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 85///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137936///dataset: EXPO///series: EXPO///id: 184844///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137937///dataset: EXPO///series: EXPO///id: 184845///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 35///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
  samplename: GSM137939///dataset: EXPO///series: EXPO///id: 184853///er: 1///pgr: 0///her2: NA///size: NA///node: NA///age: 85///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137943///dataset: EXPO///series: EXPO///id: 187219///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 85///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137944///dataset: EXPO///series: EXPO///id: 187220///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 55///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137950///dataset: EXPO///series: EXPO///id: 187234///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137953///dataset: EXPO///series: EXPO///id: 187238///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137963///dataset: EXPO///series: EXPO///id: 191420///er: 1///pgr: 0///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137978///dataset: EXPO///series: EXPO///id: 195189///er: 1///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
  samplename: GSM137984///dataset: EXPO///series: EXPO///id: 195199///er: 0///pgr: NA///her2: 0///size: NA///node: NA///age: 55///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137987///dataset: EXPO///series: EXPO///id: 199344///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
  samplename: GSM137988///dataset: EXPO///series: EXPO///id: 199351///er: 0///pgr: 0///her2: NA///size: NA///node: NA///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137989///dataset: EXPO///series: EXPO///id: 199352///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137991///dataset: EXPO///series: EXPO///id: 199356///er: 0///pgr: 0///her2: 1///size: NA///node: NA///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM137994///dataset: EXPO///series: EXPO///id: 199360///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 95///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM138011///dataset: EXPO///series: EXPO///id: 205693///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 85///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM138014///dataset: EXPO///series: EXPO///id: 207170///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM138027///dataset: EXPO///series: EXPO///id: 215917///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 95///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM138028///dataset: EXPO///series: EXPO///id: 215927///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM138031///dataset: EXPO///series: EXPO///id: 215931///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 45///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM138035///dataset: EXPO///series: EXPO///id: 215935///er: 0///pgr: 0///her2: 1///size: NA///node: NA///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM138041///dataset: EXPO///series: EXPO///id: 215944///er: 0///pgr: 0///her2: 1///size: NA///node: NA///age: 75///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
  samplename: GSM138049///dataset: EXPO///series: EXPO///id: 231855///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 45///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM138051///dataset: EXPO///series: EXPO///id: 231868///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
    samplename: GSM138054///dataset: EXPO///series: EXPO///id: 89262///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 55///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
 samplename: GSM152564///dataset: EXPO///series: EXPO///id: 101144///er: 1///pgr: 0///her2: NA///size: NA///node: NA///age: 75///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
 samplename: GSM152569///dataset: EXPO///series: EXPO///id: 1234///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 75///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM152570///dataset: EXPO///series: EXPO///id: 127820///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 55///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM152571///dataset: EXPO///series: EXPO///id: 127826///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM152576///dataset: EXPO///series: EXPO///id: 127841///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 45///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM152585///dataset: EXPO///series: EXPO///id: 133639///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 75///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
  samplename: GSM152590///dataset: EXPO///series: EXPO///id: 151607///er: 1///pgr: 0///her2: NA///size: NA///node: NA///age: 85///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM152605///dataset: EXPO///series: EXPO///id: 184847///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM152615///dataset: EXPO///series: EXPO///id: 193022///er: 1///pgr: 1///her2: 1///size: NA///node: NA///age: 45///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM152616///dataset: EXPO///series: EXPO///id: 193024///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM152618///dataset: EXPO///series: EXPO///id: 193033///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM152620///dataset: EXPO///series: EXPO///id: 193035///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 65///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM152621///dataset: EXPO///series: EXPO///id: 193036///er: 1///pgr: 0///her2: 0///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM152627///dataset: EXPO///series: EXPO///id: 199373///er: 0///pgr: 0///her2: 0///size: NA///node: NA///age: 55///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
   samplename: GSM152645///dataset: EXPO///series: EXPO///id: 205705///er: 1///pgr: 1///her2: 0///size: NA///node: NA///age: 55///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
samplename: GSM152661///dataset: EXPO///series: EXPO///id: 211394///er: NA///pgr: NA///her2: NA///size: NA///node: NA///age: 65///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: NA 
                                                                                                                                                                                                                                             1 
                                                                                                                                                                                                                                       (Other) 
                                                                                                                                                                                                                                           254 

duplicates: 
EXPO.EXPO_GSM53027 EXPO.EXPO_GSM53059               NA's 
                 1                  1                351 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE2109 }
\keyword{datasets}
