\name{ FNCLCC_eset }
\alias{ FNCLCC_eset }
\docType{data}
\title{ FNCLCC_eset }
\usage{data( FNCLCC_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/?term=17659439 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE7017 
  PMIDs: 17659439 
  No abstract available.
  notes:
   summary:     
      A potentially more powerful clinicogenomic model was created by combining a subset of relevant genes from an already publ
ished gene expression signature and a commonly used clinical prognostic model (NPI). The genes in this model are known to
 have a role in breast cancer, carcinogenesis, or chemotherapy resistance.
   version:     
      2015-04-27 19:07:20
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: UMGC_00005 UMGC_00007 ... UMGC_09018 (6064 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 6064 features, 150 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
FNCLCC_FEC100_0001_A FNCLCC_FEC100_0002_A FNCLCC_FEC100_0003_A FNCLCC_FEC100_0004_A FNCLCC_FEC100_0005_A FNCLCC_FEC100_0006_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0007_A FNCLCC_FEC100_0008_A FNCLCC_FEC100_0009_A FNCLCC_FEC100_0010_A FNCLCC_FEC100_0011_A FNCLCC_FEC100_0012_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0013_A FNCLCC_FEC100_0014_A FNCLCC_FEC100_0015_A FNCLCC_FEC100_0016_A FNCLCC_FEC100_0017_A FNCLCC_FEC100_0018_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0019_A FNCLCC_FEC100_0020_A FNCLCC_FEC100_0021_A FNCLCC_FEC100_0022_A FNCLCC_FEC100_0023_A FNCLCC_FEC100_0024_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0025_A FNCLCC_FEC100_0026_A FNCLCC_FEC100_0027_A FNCLCC_FEC100_0028_A FNCLCC_FEC100_0029_A FNCLCC_FEC100_0030_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0031_A FNCLCC_FEC100_0032_A FNCLCC_FEC100_0033_A FNCLCC_FEC100_0034_A FNCLCC_FEC100_0035_A FNCLCC_FEC100_0036_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0037_A FNCLCC_FEC100_0038_A FNCLCC_FEC100_0039_A FNCLCC_FEC100_0040_A FNCLCC_FEC100_0041_A FNCLCC_FEC100_0042_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0043_A FNCLCC_FEC100_0044_A FNCLCC_FEC100_0045_A FNCLCC_FEC100_0046_A FNCLCC_FEC100_0047_A FNCLCC_FEC100_0048_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0049_A FNCLCC_FEC100_0050_A FNCLCC_FEC100_0051_A FNCLCC_FEC100_0052_A FNCLCC_FEC100_0053_A FNCLCC_FEC100_0054_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0055_A FNCLCC_FEC100_0056_A FNCLCC_FEC100_0057_A FNCLCC_FEC100_0058_A FNCLCC_FEC100_0059_A FNCLCC_FEC100_0060_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0061_A FNCLCC_FEC100_0062_A FNCLCC_FEC100_0063_A FNCLCC_FEC100_0064_A FNCLCC_FEC100_0065_A FNCLCC_FEC100_0066_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0067_A FNCLCC_FEC100_0068_A FNCLCC_FEC100_0069_A FNCLCC_FEC100_0070_A FNCLCC_FEC100_0071_A FNCLCC_FEC100_0072_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0073_A FNCLCC_FEC100_0074_A FNCLCC_FEC100_0075_A FNCLCC_FEC100_0076_A FNCLCC_FEC100_0077_A FNCLCC_FEC100_0078_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0079_A FNCLCC_FEC100_0080_A FNCLCC_FEC100_0081_A FNCLCC_FEC100_0082_A FNCLCC_FEC100_0083_A FNCLCC_FEC100_0084_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0085_A FNCLCC_FEC100_0086_A FNCLCC_FEC100_0087_A FNCLCC_FEC100_0088_A FNCLCC_FEC100_0089_A FNCLCC_FEC100_0090_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0091_A FNCLCC_FEC100_0092_A FNCLCC_FEC100_0093_A FNCLCC_FEC100_0094_A FNCLCC_FEC100_0095_A FNCLCC_FEC100_0096_A 
                   1                    1                    1                    1                    1                    1 
FNCLCC_FEC100_0097_A FNCLCC_FEC100_0098_A FNCLCC_FEC100_0099_A              (Other) 
                   1                    1                    1                   51 

alt_sample_name: 
FEC100_0001_A FEC100_0002_A FEC100_0003_A FEC100_0004_A FEC100_0005_A FEC100_0006_A FEC100_0007_A FEC100_0008_A FEC100_0009_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0010_A FEC100_0011_A FEC100_0012_A FEC100_0013_A FEC100_0014_A FEC100_0015_A FEC100_0016_A FEC100_0017_A FEC100_0018_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0019_A FEC100_0020_A FEC100_0021_A FEC100_0022_A FEC100_0023_A FEC100_0024_A FEC100_0025_A FEC100_0026_A FEC100_0027_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0028_A FEC100_0029_A FEC100_0030_A FEC100_0031_A FEC100_0032_A FEC100_0033_A FEC100_0034_A FEC100_0035_A FEC100_0036_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0037_A FEC100_0038_A FEC100_0039_A FEC100_0040_A FEC100_0041_A FEC100_0042_A FEC100_0043_A FEC100_0044_A FEC100_0045_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0046_A FEC100_0047_A FEC100_0048_A FEC100_0049_A FEC100_0050_A FEC100_0051_A FEC100_0052_A FEC100_0053_A FEC100_0054_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0055_A FEC100_0056_A FEC100_0057_A FEC100_0058_A FEC100_0059_A FEC100_0060_A FEC100_0061_A FEC100_0062_A FEC100_0063_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0064_A FEC100_0065_A FEC100_0066_A FEC100_0067_A FEC100_0068_A FEC100_0069_A FEC100_0070_A FEC100_0071_A FEC100_0072_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0073_A FEC100_0074_A FEC100_0075_A FEC100_0076_A FEC100_0077_A FEC100_0078_A FEC100_0079_A FEC100_0080_A FEC100_0081_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0082_A FEC100_0083_A FEC100_0084_A FEC100_0085_A FEC100_0086_A FEC100_0087_A FEC100_0088_A FEC100_0089_A FEC100_0090_A 
            1             1             1             1             1             1             1             1             1 
FEC100_0091_A FEC100_0092_A FEC100_0093_A FEC100_0094_A FEC100_0095_A FEC100_0096_A FEC100_0097_A FEC100_0098_A FEC100_0099_A 
            1             1             1             1             1             1             1             1             1 
      (Other) 
           51 

sample_type: 
tumor 
  150 

N: 
  1 
150 

treatment: 
chemotherapy 
         150 

batch: 
FNCLCC 
   150 

uncurated_author_metadata: 
samplename: FNCLCC_FEC100_0001_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0001_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0002_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0002_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0003_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0003_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0004_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0004_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0005_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0005_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0006_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0006_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0007_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0007_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0008_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0008_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0009_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0009_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0010_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0010_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0011_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0011_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0012_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0012_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0013_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0013_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0014_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0014_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0015_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0015_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0016_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0016_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0017_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0017_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0018_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0018_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0019_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0019_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0020_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0020_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0021_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0021_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0022_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0022_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0023_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0023_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0024_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0024_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0025_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0025_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0026_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0026_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0027_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0027_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0028_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0028_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0029_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0029_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0030_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0030_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0031_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0031_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0032_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0032_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0033_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0033_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0034_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0034_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0035_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0035_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0036_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0036_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0037_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0037_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0038_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0038_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0039_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0039_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0040_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0040_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0041_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0041_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0042_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0042_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0043_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0043_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0044_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0044_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0045_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0045_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0046_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0046_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0047_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0047_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0048_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0048_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0049_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0049_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0050_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0050_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0051_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0051_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0052_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0052_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0053_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0053_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0054_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0054_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0055_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0055_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0056_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0056_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0057_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0057_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0058_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0058_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0059_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0059_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0060_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0060_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0061_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0061_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0062_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0062_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0063_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0063_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0064_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0064_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0065_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0065_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0066_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0066_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0067_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0067_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0068_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0068_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0069_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0069_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0070_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0070_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0071_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0071_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0072_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0072_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0073_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0073_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0074_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0074_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0075_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0075_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0076_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0076_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0077_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0077_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0078_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0078_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0079_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0079_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0080_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0080_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0081_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0081_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0082_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0082_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0083_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0083_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0084_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0084_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0085_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0085_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0086_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0086_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0087_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0087_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0088_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0088_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0089_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0089_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0090_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0090_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0091_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0091_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0092_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0092_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0093_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0093_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0094_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0094_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0095_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0095_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0096_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0096_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0097_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0097_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0098_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0098_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
samplename: FNCLCC_FEC100_0099_A///dataset: FNCLCC///series: FNCLCC///id: FEC100_0099_A///er: NA///pgr: NA///her2: NA///size: NA///node: 1///age: NA///grade: NA///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 1 
                                                                                                                                                                                                                                                                  1 
                                                                                                                                                                                                                                                            (Other) 
                                                                                                                                                                                                                                                                 51 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE7017 }
\keyword{datasets}
