\name{ IRB_eset }
\alias{ IRB_eset }
\docType{data}
\title{ IRB_eset }
\usage{data( IRB_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www-ncbi-nlm-nih-gov.proxy.wexler.hunter.cuny.edu/pubmed/?term=18498629 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6532 
  PMIDs: 18498629 
  No abstract available.
  notes:
   summary:     
      A gene classifier was identified as a predictor of clinical outcome in tamoxifen-treated breast cancer. 
   version:     
      2015-04-27 19:13:07
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 1007_s_at 1053_at ... AFFX-HUMISGF3A/M97935_MB_at (42447 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 42447 features, 129 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
IRB_GSM124994 IRB_GSM124995 IRB_GSM124996 IRB_GSM124997 IRB_GSM124998 IRB_GSM124999 IRB_GSM125000 IRB_GSM125001 IRB_GSM125002 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125003 IRB_GSM125004 IRB_GSM125005 IRB_GSM125006 IRB_GSM125007 IRB_GSM125008 IRB_GSM125009 IRB_GSM125010 IRB_GSM125011 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125012 IRB_GSM125013 IRB_GSM125014 IRB_GSM125015 IRB_GSM125016 IRB_GSM125017 IRB_GSM125018 IRB_GSM125019 IRB_GSM125020 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125021 IRB_GSM125022 IRB_GSM125023 IRB_GSM125024 IRB_GSM125025 IRB_GSM125026 IRB_GSM125027 IRB_GSM125028 IRB_GSM125029 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125030 IRB_GSM125031 IRB_GSM125032 IRB_GSM125033 IRB_GSM125034 IRB_GSM125035 IRB_GSM125036 IRB_GSM125037 IRB_GSM125038 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125039 IRB_GSM125040 IRB_GSM125041 IRB_GSM125042 IRB_GSM125043 IRB_GSM125044 IRB_GSM125045 IRB_GSM125046 IRB_GSM125047 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125048 IRB_GSM125049 IRB_GSM125050 IRB_GSM125051 IRB_GSM125052 IRB_GSM125053 IRB_GSM125054 IRB_GSM125055 IRB_GSM125056 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125057 IRB_GSM125058 IRB_GSM125059 IRB_GSM125060 IRB_GSM125061 IRB_GSM125062 IRB_GSM125063 IRB_GSM125064 IRB_GSM125065 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125066 IRB_GSM125067 IRB_GSM125068 IRB_GSM125069 IRB_GSM125070 IRB_GSM125071 IRB_GSM125072 IRB_GSM125073 IRB_GSM125074 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125075 IRB_GSM125076 IRB_GSM125077 IRB_GSM125078 IRB_GSM125079 IRB_GSM125080 IRB_GSM125081 IRB_GSM125082 IRB_GSM125083 
            1             1             1             1             1             1             1             1             1 
IRB_GSM125084 IRB_GSM125085 IRB_GSM125086 IRB_GSM125087 IRB_GSM125088 IRB_GSM125089 IRB_GSM125090 IRB_GSM125091 IRB_GSM125092 
            1             1             1             1             1             1             1             1             1 
      (Other) 
           30 

alt_sample_name: 
GSM124994 GSM124995 GSM124996 GSM124997 GSM124998 GSM124999 GSM125000 GSM125001 GSM125002 GSM125003 GSM125004 GSM125005 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM125006 GSM125007 GSM125008 GSM125009 GSM125010 GSM125011 GSM125012 GSM125013 GSM125014 GSM125015 GSM125016 GSM125017 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM125018 GSM125019 GSM125020 GSM125021 GSM125022 GSM125023 GSM125024 GSM125025 GSM125026 GSM125027 GSM125028 GSM125029 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM125030 GSM125031 GSM125032 GSM125033 GSM125034 GSM125035 GSM125036 GSM125037 GSM125038 GSM125039 GSM125040 GSM125041 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM125042 GSM125043 GSM125044 GSM125045 GSM125046 GSM125047 GSM125048 GSM125049 GSM125050 GSM125051 GSM125052 GSM125053 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM125054 GSM125055 GSM125056 GSM125057 GSM125058 GSM125059 GSM125060 GSM125061 GSM125062 GSM125063 GSM125064 GSM125065 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM125066 GSM125067 GSM125068 GSM125069 GSM125070 GSM125071 GSM125072 GSM125073 GSM125074 GSM125075 GSM125076 GSM125077 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM125078 GSM125079 GSM125080 GSM125081 GSM125082 GSM125083 GSM125084 GSM125085 GSM125086 GSM125087 GSM125088 GSM125089 
        1         1         1         1         1         1         1         1         1         1         1         1 
GSM125090 GSM125091 GSM125092   (Other) 
        1         1         1        30 

sample_type: 
tumor 
  129 

er: 
negative positive 
      53       76 

her2: 
negative positive 
      98       31 

tumor_size: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  0.800   1.500   2.200   2.488   3.000   8.500 

N: 
 0  1 
64 65 

grade: 
 1  2  3 
27 32 70 

treatment: 
untreated 
      129 

batch: 
IRB 
129 

uncurated_author_metadata: 
samplename: IRB_GSM124994///dataset: IRB///series: IRB///id: GSM124994///er: 0///pgr: NA///her2: 1///size: 2.3///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM124995///dataset: IRB///series: IRB///id: GSM124995///er: 0///pgr: NA///her2: 0///size: 0.9///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM124996///dataset: IRB///series: IRB///id: GSM124996///er: 0///pgr: NA///her2: 1///size: 2.5///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM124997///dataset: IRB///series: IRB///id: GSM124997///er: 0///pgr: NA///her2: 0///size: 3.5///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM124998///dataset: IRB///series: IRB///id: GSM124998///er: 0///pgr: NA///her2: 0///size: 2.1///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM124999///dataset: IRB///series: IRB///id: GSM124999///er: 0///pgr: NA///her2: 1///size: 2.5///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125000///dataset: IRB///series: IRB///id: GSM125000///er: 0///pgr: NA///her2: 0///size: 1.2///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125001///dataset: IRB///series: IRB///id: GSM125001///er: 0///pgr: NA///her2: 1///size: 2.1///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125002///dataset: IRB///series: IRB///id: GSM125002///er: 0///pgr: NA///her2: 0///size: 2.5///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125003///dataset: IRB///series: IRB///id: GSM125003///er: 0///pgr: NA///her2: 0///size: 3.3///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125004///dataset: IRB///series: IRB///id: GSM125004///er: 0///pgr: NA///her2: 0///size: 2.5///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125005///dataset: IRB///series: IRB///id: GSM125005///er: 0///pgr: NA///her2: 0///size: 3///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125006///dataset: IRB///series: IRB///id: GSM125006///er: 0///pgr: NA///her2: 0///size: 1.8///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125007///dataset: IRB///series: IRB///id: GSM125007///er: 0///pgr: NA///her2: 0///size: 4///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125008///dataset: IRB///series: IRB///id: GSM125008///er: 0///pgr: NA///her2: 0///size: 2.1///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125009///dataset: IRB///series: IRB///id: GSM125009///er: 0///pgr: NA///her2: 0///size: 1///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125010///dataset: IRB///series: IRB///id: GSM125010///er: 0///pgr: NA///her2: 0///size: 0.9///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125011///dataset: IRB///series: IRB///id: GSM125011///er: 0///pgr: NA///her2: 0///size: 2.2///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125012///dataset: IRB///series: IRB///id: GSM125012///er: 0///pgr: NA///her2: 0///size: 1///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125013///dataset: IRB///series: IRB///id: GSM125013///er: 0///pgr: NA///her2: 1///size: 3.5///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125014///dataset: IRB///series: IRB///id: GSM125014///er: 0///pgr: NA///her2: 0///size: 2.5///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125015///dataset: IRB///series: IRB///id: GSM125015///er: 0///pgr: NA///her2: 0///size: 2.6///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125016///dataset: IRB///series: IRB///id: GSM125016///er: 0///pgr: NA///her2: 0///size: 3///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125017///dataset: IRB///series: IRB///id: GSM125017///er: 0///pgr: NA///her2: 0///size: 1.3///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125018///dataset: IRB///series: IRB///id: GSM125018///er: 0///pgr: NA///her2: 0///size: 1.4///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125019///dataset: IRB///series: IRB///id: GSM125019///er: 0///pgr: NA///her2: 0///size: 1.7///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125020///dataset: IRB///series: IRB///id: GSM125020///er: 0///pgr: NA///her2: 0///size: 2.3///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125021///dataset: IRB///series: IRB///id: GSM125021///er: 0///pgr: NA///her2: 1///size: 3.4///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125022///dataset: IRB///series: IRB///id: GSM125022///er: 0///pgr: NA///her2: 0///size: 1.5///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125023///dataset: IRB///series: IRB///id: GSM125023///er: 0///pgr: NA///her2: 0///size: 2.5///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125024///dataset: IRB///series: IRB///id: GSM125024///er: 0///pgr: NA///her2: 0///size: 2.8///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125025///dataset: IRB///series: IRB///id: GSM125025///er: 0///pgr: NA///her2: 0///size: 2.2///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125026///dataset: IRB///series: IRB///id: GSM125026///er: 1///pgr: NA///her2: 0///size: 3.5///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125027///dataset: IRB///series: IRB///id: GSM125027///er: 1///pgr: NA///her2: 1///size: 1///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125028///dataset: IRB///series: IRB///id: GSM125028///er: 1///pgr: NA///her2: 0///size: 1.7///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125029///dataset: IRB///series: IRB///id: GSM125029///er: 1///pgr: NA///her2: 1///size: 1.4///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125030///dataset: IRB///series: IRB///id: GSM125030///er: 1///pgr: NA///her2: 0///size: 2.6///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125031///dataset: IRB///series: IRB///id: GSM125031///er: 1///pgr: NA///her2: 0///size: 0.9///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125032///dataset: IRB///series: IRB///id: GSM125032///er: 1///pgr: NA///her2: 0///size: 0.8///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125033///dataset: IRB///series: IRB///id: GSM125033///er: 1///pgr: NA///her2: 1///size: 2.2///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125034///dataset: IRB///series: IRB///id: GSM125034///er: 0///pgr: NA///her2: 1///size: 1.8///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125035///dataset: IRB///series: IRB///id: GSM125035///er: 1///pgr: NA///her2: 0///size: 2.2///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125036///dataset: IRB///series: IRB///id: GSM125036///er: 1///pgr: NA///her2: 1///size: 4.2///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125037///dataset: IRB///series: IRB///id: GSM125037///er: 0///pgr: NA///her2: 1///size: 2.5///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125038///dataset: IRB///series: IRB///id: GSM125038///er: 1///pgr: NA///her2: 0///size: 2.3///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125039///dataset: IRB///series: IRB///id: GSM125039///er: 1///pgr: NA///her2: 1///size: 2///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125040///dataset: IRB///series: IRB///id: GSM125040///er: 1///pgr: NA///her2: 1///size: 1.7///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125041///dataset: IRB///series: IRB///id: GSM125041///er: 1///pgr: NA///her2: 0///size: 1.1///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125042///dataset: IRB///series: IRB///id: GSM125042///er: 1///pgr: NA///her2: 0///size: 2.5///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125043///dataset: IRB///series: IRB///id: GSM125043///er: 1///pgr: NA///her2: 0///size: 0.9///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125044///dataset: IRB///series: IRB///id: GSM125044///er: 0///pgr: NA///her2: 0///size: 3///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125045///dataset: IRB///series: IRB///id: GSM125045///er: 1///pgr: NA///her2: 0///size: 2.3///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125046///dataset: IRB///series: IRB///id: GSM125046///er: 1///pgr: NA///her2: 0///size: 1.5///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125047///dataset: IRB///series: IRB///id: GSM125047///er: 1///pgr: NA///her2: 0///size: 4///node: 1///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125048///dataset: IRB///series: IRB///id: GSM125048///er: 0///pgr: NA///her2: 1///size: 1.5///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125049///dataset: IRB///series: IRB///id: GSM125049///er: 1///pgr: NA///her2: 0///size: 1.1///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125050///dataset: IRB///series: IRB///id: GSM125050///er: 1///pgr: NA///her2: 0///size: 5.5///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125051///dataset: IRB///series: IRB///id: GSM125051///er: 1///pgr: NA///her2: 0///size: 2.3///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125052///dataset: IRB///series: IRB///id: GSM125052///er: 1///pgr: NA///her2: 0///size: 1///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125053///dataset: IRB///series: IRB///id: GSM125053///er: 0///pgr: NA///her2: 0///size: 1.1///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125054///dataset: IRB///series: IRB///id: GSM125054///er: 0///pgr: NA///her2: 0///size: 1.1///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125055///dataset: IRB///series: IRB///id: GSM125055///er: 1///pgr: NA///her2: 0///size: 2.5///node: 1///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125056///dataset: IRB///series: IRB///id: GSM125056///er: 0///pgr: NA///her2: 0///size: 3///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125057///dataset: IRB///series: IRB///id: GSM125057///er: 0///pgr: NA///her2: 0///size: 5.2///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125058///dataset: IRB///series: IRB///id: GSM125058///er: 0///pgr: NA///her2: 0///size: 2.5///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125059///dataset: IRB///series: IRB///id: GSM125059///er: 1///pgr: NA///her2: 0///size: 2.5///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125060///dataset: IRB///series: IRB///id: GSM125060///er: 1///pgr: NA///her2: 0///size: 1.1///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125061///dataset: IRB///series: IRB///id: GSM125061///er: 1///pgr: NA///her2: 0///size: 1.7///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125062///dataset: IRB///series: IRB///id: GSM125062///er: 1///pgr: NA///her2: 0///size: 2.1///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125063///dataset: IRB///series: IRB///id: GSM125063///er: 1///pgr: NA///her2: 0///size: 1.5///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125064///dataset: IRB///series: IRB///id: GSM125064///er: 0///pgr: NA///her2: 0///size: 1.9///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125065///dataset: IRB///series: IRB///id: GSM125065///er: 1///pgr: NA///her2: 0///size: 3.5///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125066///dataset: IRB///series: IRB///id: GSM125066///er: 0///pgr: NA///her2: 1///size: 2.2///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125067///dataset: IRB///series: IRB///id: GSM125067///er: 1///pgr: NA///her2: 0///size: 1.5///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125068///dataset: IRB///series: IRB///id: GSM125068///er: 0///pgr: NA///her2: 0///size: 2///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125069///dataset: IRB///series: IRB///id: GSM125069///er: 1///pgr: NA///her2: 1///size: 1.3///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125070///dataset: IRB///series: IRB///id: GSM125070///er: 0///pgr: NA///her2: 1///size: 2.5///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125071///dataset: IRB///series: IRB///id: GSM125071///er: 0///pgr: NA///her2: 0///size: 1.7///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125072///dataset: IRB///series: IRB///id: GSM125072///er: 1///pgr: NA///her2: 1///size: 2.9///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125073///dataset: IRB///series: IRB///id: GSM125073///er: 0///pgr: NA///her2: 1///size: 3///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125074///dataset: IRB///series: IRB///id: GSM125074///er: 1///pgr: NA///her2: 0///size: 1.1///node: 1///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125075///dataset: IRB///series: IRB///id: GSM125075///er: 1///pgr: NA///her2: 0///size: 2///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125076///dataset: IRB///series: IRB///id: GSM125076///er: 1///pgr: NA///her2: 1///size: 1.2///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125077///dataset: IRB///series: IRB///id: GSM125077///er: 1///pgr: NA///her2: 0///size: 3.5///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125078///dataset: IRB///series: IRB///id: GSM125078///er: 1///pgr: NA///her2: 1///size: 3.1///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125079///dataset: IRB///series: IRB///id: GSM125079///er: 1///pgr: NA///her2: 0///size: 1.2///node: 1///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125080///dataset: IRB///series: IRB///id: GSM125080///er: 1///pgr: NA///her2: 0///size: 4.2///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125081///dataset: IRB///series: IRB///id: GSM125081///er: 1///pgr: NA///her2: 0///size: 3///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125082///dataset: IRB///series: IRB///id: GSM125082///er: 1///pgr: NA///her2: 0///size: 2.2///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125083///dataset: IRB///series: IRB///id: GSM125083///er: 1///pgr: NA///her2: 0///size: 1.3///node: 1///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125084///dataset: IRB///series: IRB///id: GSM125084///er: 1///pgr: NA///her2: 0///size: 1.2///node: 1///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
  samplename: IRB_GSM125085///dataset: IRB///series: IRB///id: GSM125085///er: 1///pgr: NA///her2: 1///size: 2///node: 0///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125086///dataset: IRB///series: IRB///id: GSM125086///er: 1///pgr: NA///her2: 0///size: 1.1///node: 0///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125087///dataset: IRB///series: IRB///id: GSM125087///er: 0///pgr: NA///her2: 1///size: 2.9///node: 1///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125088///dataset: IRB///series: IRB///id: GSM125088///er: 1///pgr: NA///her2: 0///size: 2.5///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125089///dataset: IRB///series: IRB///id: GSM125089///er: 1///pgr: NA///her2: 0///size: 2.5///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125090///dataset: IRB///series: IRB///id: GSM125090///er: 1///pgr: NA///her2: 1///size: 4.5///node: 1///age: NA///grade: 3///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125091///dataset: IRB///series: IRB///id: GSM125091///er: 1///pgr: NA///her2: 0///size: 1.2///node: 0///age: NA///grade: 2///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
samplename: IRB_GSM125092///dataset: IRB///series: IRB///id: GSM125092///er: 1///pgr: NA///her2: 0///size: 1.5///node: 1///age: NA///grade: 1///t.dmfs: NA///e.dmfs: NA///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                               1 
                                                                                                                                                                                                                                         (Other) 
                                                                                                                                                                                                                                              30 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE6532 }
\keyword{datasets}
