\name{ MAINZ_eset }
\alias{ MAINZ_eset }
\docType{data}
\title{ MAINZ_eset }
\usage{data( MAINZ_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/?term=18593943 
  Title:  
  URL: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE11121 
  PMIDs: 18593943 
  No abstract available.
  notes:
   summary:     
      Poor prognosis is noted in tumors with low ER expression, showing the highest level of proliferative activity. In some tu
mors with highly expressed B-cell or T-cell metagenes, metastases rarely occurred, even with high proliferation and low E
R expression.
   version:     
      2015-04-27 19:16:04
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 1007_s_at 1053_at ... AFFX-HUMISGF3A/M97935_MB_at (20967 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 20967 features, 200 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
MAINZ_BC6001 MAINZ_BC6002 MAINZ_BC6003 MAINZ_BC6004 MAINZ_BC6005 MAINZ_BC6006 MAINZ_BC6007 MAINZ_BC6008 MAINZ_BC6009 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6011 MAINZ_BC6012 MAINZ_BC6013 MAINZ_BC6014 MAINZ_BC6015 MAINZ_BC6016 MAINZ_BC6017 MAINZ_BC6019 MAINZ_BC6020 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6021 MAINZ_BC6022 MAINZ_BC6024 MAINZ_BC6025 MAINZ_BC6026 MAINZ_BC6027 MAINZ_BC6028 MAINZ_BC6029 MAINZ_BC6030 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6031 MAINZ_BC6032 MAINZ_BC6033 MAINZ_BC6035 MAINZ_BC6036 MAINZ_BC6037 MAINZ_BC6039 MAINZ_BC6040 MAINZ_BC6041 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6042 MAINZ_BC6043 MAINZ_BC6044 MAINZ_BC6045 MAINZ_BC6046 MAINZ_BC6047 MAINZ_BC6048 MAINZ_BC6049 MAINZ_BC6050 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6051 MAINZ_BC6052 MAINZ_BC6053 MAINZ_BC6054 MAINZ_BC6055 MAINZ_BC6056 MAINZ_BC6057 MAINZ_BC6058 MAINZ_BC6059 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6060 MAINZ_BC6061 MAINZ_BC6062 MAINZ_BC6063 MAINZ_BC6064 MAINZ_BC6065 MAINZ_BC6066 MAINZ_BC6067 MAINZ_BC6068 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6069 MAINZ_BC6070 MAINZ_BC6071 MAINZ_BC6072 MAINZ_BC6073 MAINZ_BC6074 MAINZ_BC6075 MAINZ_BC6076 MAINZ_BC6077 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6078 MAINZ_BC6079 MAINZ_BC6080 MAINZ_BC6081 MAINZ_BC6082 MAINZ_BC6083 MAINZ_BC6084 MAINZ_BC6085 MAINZ_BC6086 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6087 MAINZ_BC6088 MAINZ_BC6089 MAINZ_BC6091 MAINZ_BC6092 MAINZ_BC6093 MAINZ_BC6094 MAINZ_BC6095 MAINZ_BC6096 
           1            1            1            1            1            1            1            1            1 
MAINZ_BC6097 MAINZ_BC6098 MAINZ_BC6099 MAINZ_BC6100 MAINZ_BC6101 MAINZ_BC6102 MAINZ_BC6103 MAINZ_BC6104 MAINZ_BC6105 
           1            1            1            1            1            1            1            1            1 
     (Other) 
         101 

alt_sample_name: 
 BC6001_r: Mainz_N0_untreated_frozen _001  BC6002_r: Mainz_N0_untreated_frozen _002  BC6003_r: Mainz_N0_untreated_frozen _003 
                                        1                                         1                                         1 
 BC6004_r: Mainz_N0_untreated_frozen _004 BC6005wdh: Mainz_N0_untreated_frozen _005    BC6006: Mainz_N0_untreated_frozen _006 
                                        1                                         1                                         1 
   BC6007: Mainz_N0_untreated_frozen _007    BC6008: Mainz_N0_untreated_frozen _008    BC6009: Mainz_N0_untreated_frozen _009 
                                        1                                         1                                         1 
   BC6011: Mainz_N0_untreated_frozen _010    BC6012: Mainz_N0_untreated_frozen _011    BC6013: Mainz_N0_untreated_frozen _012 
                                        1                                         1                                         1 
   BC6014: Mainz_N0_untreated_frozen _013    BC6015: Mainz_N0_untreated_frozen _014    BC6016: Mainz_N0_untreated_frozen _015 
                                        1                                         1                                         1 
   BC6017: Mainz_N0_untreated_frozen _016    BC6019: Mainz_N0_untreated_frozen _017  BC6020_3: Mainz_N0_untreated_frozen _018 
                                        1                                         1                                         1 
   BC6021: Mainz_N0_untreated_frozen _019    BC6022: Mainz_N0_untreated_frozen _020    BC6024: Mainz_N0_untreated_frozen _021 
                                        1                                         1                                         1 
   BC6025: Mainz_N0_untreated_frozen _022    BC6026: Mainz_N0_untreated_frozen _023    BC6027: Mainz_N0_untreated_frozen _024 
                                        1                                         1                                         1 
 BC6028_3: Mainz_N0_untreated_frozen _025    BC6029: Mainz_N0_untreated_frozen _026    BC6030: Mainz_N0_untreated_frozen _027 
                                        1                                         1                                         1 
   BC6031: Mainz_N0_untreated_frozen _028    BC6032: Mainz_N0_untreated_frozen _029    BC6033: Mainz_N0_untreated_frozen _030 
                                        1                                         1                                         1 
   BC6035: Mainz_N0_untreated_frozen _031    BC6036: Mainz_N0_untreated_frozen _032    BC6037: Mainz_N0_untreated_frozen _033 
                                        1                                         1                                         1 
   BC6039: Mainz_N0_untreated_frozen _034    BC6040: Mainz_N0_untreated_frozen _035    BC6041: Mainz_N0_untreated_frozen _036 
                                        1                                         1                                         1 
   BC6042: Mainz_N0_untreated_frozen _037    BC6043: Mainz_N0_untreated_frozen _038    BC6044: Mainz_N0_untreated_frozen _039 
                                        1                                         1                                         1 
   BC6045: Mainz_N0_untreated_frozen _040    BC6046: Mainz_N0_untreated_frozen _041    BC6047: Mainz_N0_untreated_frozen _042 
                                        1                                         1                                         1 
   BC6048: Mainz_N0_untreated_frozen _043    BC6049: Mainz_N0_untreated_frozen _044    BC6050: Mainz_N0_untreated_frozen _045 
                                        1                                         1                                         1 
 BC6051_2: Mainz_N0_untreated_frozen _046    BC6052: Mainz_N0_untreated_frozen _047    BC6053: Mainz_N0_untreated_frozen _048 
                                        1                                         1                                         1 
   BC6054: Mainz_N0_untreated_frozen _049    BC6055: Mainz_N0_untreated_frozen _050 BC6056wdh: Mainz_N0_untreated_frozen _051 
                                        1                                         1                                         1 
   BC6057: Mainz_N0_untreated_frozen _052    BC6058: Mainz_N0_untreated_frozen _053    BC6059: Mainz_N0_untreated_frozen _054 
                                        1                                         1                                         1 
BC6060wdh: Mainz_N0_untreated_frozen _055    BC6061: Mainz_N0_untreated_frozen _056 BC6062wdh: Mainz_N0_untreated_frozen _057 
                                        1                                         1                                         1 
   BC6063: Mainz_N0_untreated_frozen _058  BC6064_2: Mainz_N0_untreated_frozen _059    BC6065: Mainz_N0_untreated_frozen _060 
                                        1                                         1                                         1 
BC6066wdh: Mainz_N0_untreated_frozen _061    BC6067: Mainz_N0_untreated_frozen _062    BC6068: Mainz_N0_untreated_frozen _063 
                                        1                                         1                                         1 
   BC6069: Mainz_N0_untreated_frozen _064    BC6070: Mainz_N0_untreated_frozen _065    BC6071: Mainz_N0_untreated_frozen _066 
                                        1                                         1                                         1 
   BC6072: Mainz_N0_untreated_frozen _067    BC6073: Mainz_N0_untreated_frozen _068    BC6074: Mainz_N0_untreated_frozen _069 
                                        1                                         1                                         1 
   BC6075: Mainz_N0_untreated_frozen _070    BC6076: Mainz_N0_untreated_frozen _071    BC6077: Mainz_N0_untreated_frozen _072 
                                        1                                         1                                         1 
   BC6078: Mainz_N0_untreated_frozen _073    BC6079: Mainz_N0_untreated_frozen _074    BC6080: Mainz_N0_untreated_frozen _075 
                                        1                                         1                                         1 
   BC6081: Mainz_N0_untreated_frozen _076    BC6082: Mainz_N0_untreated_frozen _077    BC6083: Mainz_N0_untreated_frozen _078 
                                        1                                         1                                         1 
   BC6084: Mainz_N0_untreated_frozen _079    BC6085: Mainz_N0_untreated_frozen _080    BC6086: Mainz_N0_untreated_frozen _081 
                                        1                                         1                                         1 
   BC6087: Mainz_N0_untreated_frozen _082    BC6088: Mainz_N0_untreated_frozen _083    BC6089: Mainz_N0_untreated_frozen _084 
                                        1                                         1                                         1 
   BC6091: Mainz_N0_untreated_frozen _085    BC6092: Mainz_N0_untreated_frozen _086    BC6093: Mainz_N0_untreated_frozen _087 
                                        1                                         1                                         1 
   BC6094: Mainz_N0_untreated_frozen _088    BC6095: Mainz_N0_untreated_frozen _089    BC6096: Mainz_N0_untreated_frozen _090 
                                        1                                         1                                         1 
   BC6097: Mainz_N0_untreated_frozen _091    BC6098: Mainz_N0_untreated_frozen _092    BC6099: Mainz_N0_untreated_frozen _093 
                                        1                                         1                                         1 
   BC6100: Mainz_N0_untreated_frozen _094 BC6101wdh: Mainz_N0_untreated_frozen _095    BC6102: Mainz_N0_untreated_frozen _096 
                                        1                                         1                                         1 
   BC6103: Mainz_N0_untreated_frozen _097    BC6104: Mainz_N0_untreated_frozen _098    BC6105: Mainz_N0_untreated_frozen _099 
                                        1                                         1                                         1 
                                  (Other) 
                                      101 

sample_type: 
tumor 
  200 

er: 
negative positive 
      38      162 

tumor_size: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  0.100   1.500   2.000   2.070   2.425   6.000 

N: 
  0 
200 

age_at_initial_pathologic_diagnosis: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  25.00   50.00   60.00   59.98   69.00   90.00 

grade: 
  1   2   3 
 29 136  35 

dmfs_days: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
     30    1905    2715    2816    3855    7200 

dmfs_status: 
deceased_or_recurrence    living_norecurrence 
                    46                    154 

treatment: 
untreated 
      200 

batch: 
MAINZ 
  200 

uncurated_author_metadata: 
 samplename: MAINZ_BC6001///dataset: MAINZ///series: MAINZ///id: BC6001_r: Mainz_N0_untreated_frozen _001///er: 0///pgr: NA///her2: NA///size: 1.8///node: 0///age: 54///grade: 2///t.dmfs: 2760///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
 samplename: MAINZ_BC6002///dataset: MAINZ///series: MAINZ///id: BC6002_r: Mainz_N0_untreated_frozen _002///er: 1///pgr: NA///her2: NA///size: 2.5///node: 0///age: 54///grade: 3///t.dmfs: 2130///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
 samplename: MAINZ_BC6003///dataset: MAINZ///series: MAINZ///id: BC6003_r: Mainz_N0_untreated_frozen _003///er: 1///pgr: NA///her2: NA///size: 1.5///node: 0///age: 42///grade: 3///t.dmfs: 1740///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
 samplename: MAINZ_BC6004///dataset: MAINZ///series: MAINZ///id: BC6004_r: Mainz_N0_untreated_frozen _004///er: 1///pgr: NA///her2: NA///size: 1.2///node: 0///age: 71///grade: 2///t.dmfs: 2040///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
samplename: MAINZ_BC6005///dataset: MAINZ///series: MAINZ///id: BC6005wdh: Mainz_N0_untreated_frozen _005///er: 1///pgr: NA///her2: NA///size: 2.4///node: 0///age: 58///grade: 2///t.dmfs: 3090///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6006///dataset: MAINZ///series: MAINZ///id: BC6006: Mainz_N0_untreated_frozen _006///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 61///grade: 2///t.dmfs: 2790///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6007///dataset: MAINZ///series: MAINZ///id: BC6007: Mainz_N0_untreated_frozen _007///er: 1///pgr: NA///her2: NA///size: 1.4///node: 0///age: 60///grade: 2///t.dmfs: 3330///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6008///dataset: MAINZ///series: MAINZ///id: BC6008: Mainz_N0_untreated_frozen _008///er: 1///pgr: NA///her2: NA///size: 1.3///node: 0///age: 64///grade: 3///t.dmfs: 2550///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6009///dataset: MAINZ///series: MAINZ///id: BC6009: Mainz_N0_untreated_frozen _009///er: 1///pgr: NA///her2: NA///size: 0.8///node: 0///age: 61///grade: 2///t.dmfs: 2940///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6011///dataset: MAINZ///series: MAINZ///id: BC6011: Mainz_N0_untreated_frozen _010///er: 1///pgr: NA///her2: NA///size: 1.6///node: 0///age: 72///grade: 1///t.dmfs: 2430///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6012///dataset: MAINZ///series: MAINZ///id: BC6012: Mainz_N0_untreated_frozen _011///er: 1///pgr: NA///her2: NA///size: 1.4///node: 0///age: 56///grade: 2///t.dmfs: 2370///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6013///dataset: MAINZ///series: MAINZ///id: BC6013: Mainz_N0_untreated_frozen _012///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 44///grade: 2///t.dmfs: 2400///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6014///dataset: MAINZ///series: MAINZ///id: BC6014: Mainz_N0_untreated_frozen _013///er: 0///pgr: NA///her2: NA///size: 2.2///node: 0///age: 57///grade: 2///t.dmfs: 2670///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6015///dataset: MAINZ///series: MAINZ///id: BC6015: Mainz_N0_untreated_frozen _014///er: 1///pgr: NA///her2: NA///size: 2.3///node: 0///age: 69///grade: 2///t.dmfs: 2520///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
    samplename: MAINZ_BC6016///dataset: MAINZ///series: MAINZ///id: BC6016: Mainz_N0_untreated_frozen _015///er: 1///pgr: NA///her2: NA///size: 2.3///node: 0///age: 84///grade: 2///t.dmfs: 420///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6017///dataset: MAINZ///series: MAINZ///id: BC6017: Mainz_N0_untreated_frozen _016///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 58///grade: 3///t.dmfs: 2640///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6019///dataset: MAINZ///series: MAINZ///id: BC6019: Mainz_N0_untreated_frozen _017///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 67///grade: 2///t.dmfs: 2400///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
 samplename: MAINZ_BC6020///dataset: MAINZ///series: MAINZ///id: BC6020_3: Mainz_N0_untreated_frozen _018///er: 1///pgr: NA///her2: NA///size: 0.9///node: 0///age: 51///grade: 2///t.dmfs: 2070///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6021///dataset: MAINZ///series: MAINZ///id: BC6021: Mainz_N0_untreated_frozen _019///er: 1///pgr: NA///her2: NA///size: 4.4///node: 0///age: 66///grade: 2///t.dmfs: 2790///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6022///dataset: MAINZ///series: MAINZ///id: BC6022: Mainz_N0_untreated_frozen _020///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 51///grade: 2///t.dmfs: 2640///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6024///dataset: MAINZ///series: MAINZ///id: BC6024: Mainz_N0_untreated_frozen _021///er: 1///pgr: NA///her2: NA///size: 1.2///node: 0///age: 50///grade: 1///t.dmfs: 2880///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6025///dataset: MAINZ///series: MAINZ///id: BC6025: Mainz_N0_untreated_frozen _022///er: 1///pgr: NA///her2: NA///size: 1.6///node: 0///age: 78///grade: 2///t.dmfs: 2340///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6026///dataset: MAINZ///series: MAINZ///id: BC6026: Mainz_N0_untreated_frozen _023///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 43///grade: 2///t.dmfs: 2370///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
    samplename: MAINZ_BC6027///dataset: MAINZ///series: MAINZ///id: BC6027: Mainz_N0_untreated_frozen _024///er: 1///pgr: NA///her2: NA///size: 1.3///node: 0///age: 45///grade: 2///t.dmfs: 780///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
 samplename: MAINZ_BC6028///dataset: MAINZ///series: MAINZ///id: BC6028_3: Mainz_N0_untreated_frozen _025///er: 1///pgr: NA///her2: NA///size: 1.1///node: 0///age: 53///grade: 2///t.dmfs: 2460///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
    samplename: MAINZ_BC6029///dataset: MAINZ///series: MAINZ///id: BC6029: Mainz_N0_untreated_frozen _026///er: 1///pgr: NA///her2: NA///size: 3.5///node: 0///age: 50///grade: 3///t.dmfs: 450///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6030///dataset: MAINZ///series: MAINZ///id: BC6030: Mainz_N0_untreated_frozen _027///er: 1///pgr: NA///her2: NA///size: 1.6///node: 0///age: 48///grade: 2///t.dmfs: 2160///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6031///dataset: MAINZ///series: MAINZ///id: BC6031: Mainz_N0_untreated_frozen _028///er: 1///pgr: NA///her2: NA///size: 2.2///node: 0///age: 66///grade: 3///t.dmfs: 2670///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6032///dataset: MAINZ///series: MAINZ///id: BC6032: Mainz_N0_untreated_frozen _029///er: 1///pgr: NA///her2: NA///size: 2.2///node: 0///age: 57///grade: 1///t.dmfs: 2340///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6033///dataset: MAINZ///series: MAINZ///id: BC6033: Mainz_N0_untreated_frozen _030///er: 1///pgr: NA///her2: NA///size: 1.5///node: 0///age: 71///grade: 1///t.dmfs: 2070///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
    samplename: MAINZ_BC6035///dataset: MAINZ///series: MAINZ///id: BC6035: Mainz_N0_untreated_frozen _031///er: 1///pgr: NA///her2: NA///size: 1.6///node: 0///age: 82///grade: 2///t.dmfs: 570///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6036///dataset: MAINZ///series: MAINZ///id: BC6036: Mainz_N0_untreated_frozen _032///er: 1///pgr: NA///her2: NA///size: 2.5///node: 0///age: 75///grade: 2///t.dmfs: 2190///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6037///dataset: MAINZ///series: MAINZ///id: BC6037: Mainz_N0_untreated_frozen _033///er: 1///pgr: NA///her2: NA///size: 1.5///node: 0///age: 58///grade: 2///t.dmfs: 1920///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6039///dataset: MAINZ///series: MAINZ///id: BC6039: Mainz_N0_untreated_frozen _034///er: 1///pgr: NA///her2: NA///size: 2.2///node: 0///age: 48///grade: 3///t.dmfs: 2550///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6040///dataset: MAINZ///series: MAINZ///id: BC6040: Mainz_N0_untreated_frozen _035///er: 1///pgr: NA///her2: NA///size: 1.5///node: 0///age: 51///grade: 2///t.dmfs: 2130///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6041///dataset: MAINZ///series: MAINZ///id: BC6041: Mainz_N0_untreated_frozen _036///er: 1///pgr: NA///her2: NA///size: 5.5///node: 0///age: 54///grade: 2///t.dmfs: 2250///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6042///dataset: MAINZ///series: MAINZ///id: BC6042: Mainz_N0_untreated_frozen _037///er: 0///pgr: NA///her2: NA///size: 0.1///node: 0///age: 54///grade: 2///t.dmfs: 2850///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6043///dataset: MAINZ///series: MAINZ///id: BC6043: Mainz_N0_untreated_frozen _038///er: 0///pgr: NA///her2: NA///size: 2.8///node: 0///age: 56///grade: 3///t.dmfs: 1110///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6044///dataset: MAINZ///series: MAINZ///id: BC6044: Mainz_N0_untreated_frozen _039///er: 1///pgr: NA///her2: NA///size: 1.2///node: 0///age: 59///grade: 1///t.dmfs: 2850///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6045///dataset: MAINZ///series: MAINZ///id: BC6045: Mainz_N0_untreated_frozen _040///er: 1///pgr: NA///her2: NA///size: 1.7///node: 0///age: 57///grade: 2///t.dmfs: 3330///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
    samplename: MAINZ_BC6046///dataset: MAINZ///series: MAINZ///id: BC6046: Mainz_N0_untreated_frozen _041///er: 0///pgr: NA///her2: NA///size: 1.7///node: 0///age: 68///grade: 3///t.dmfs: 510///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6047///dataset: MAINZ///series: MAINZ///id: BC6047: Mainz_N0_untreated_frozen _042///er: 1///pgr: NA///her2: NA///size: 1.2///node: 0///age: 72///grade: 2///t.dmfs: 2070///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6048///dataset: MAINZ///series: MAINZ///id: BC6048: Mainz_N0_untreated_frozen _043///er: 1///pgr: NA///her2: NA///size: 1.9///node: 0///age: 47///grade: 2///t.dmfs: 2370///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6049///dataset: MAINZ///series: MAINZ///id: BC6049: Mainz_N0_untreated_frozen _044///er: 1///pgr: NA///her2: NA///size: 1.3///node: 0///age: 64///grade: 2///t.dmfs: 2730///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6050///dataset: MAINZ///series: MAINZ///id: BC6050: Mainz_N0_untreated_frozen _045///er: 1///pgr: NA///her2: NA///size: 1.6///node: 0///age: 64///grade: 2///t.dmfs: 2160///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
 samplename: MAINZ_BC6051///dataset: MAINZ///series: MAINZ///id: BC6051_2: Mainz_N0_untreated_frozen _046///er: 1///pgr: NA///her2: NA///size: 0.3///node: 0///age: 63///grade: 3///t.dmfs: 1050///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6052///dataset: MAINZ///series: MAINZ///id: BC6052: Mainz_N0_untreated_frozen _047///er: 1///pgr: NA///her2: NA///size: 0.8///node: 0///age: 46///grade: 2///t.dmfs: 2220///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6053///dataset: MAINZ///series: MAINZ///id: BC6053: Mainz_N0_untreated_frozen _048///er: 1///pgr: NA///her2: NA///size: 1.1///node: 0///age: 61///grade: 2///t.dmfs: 3930///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6054///dataset: MAINZ///series: MAINZ///id: BC6054: Mainz_N0_untreated_frozen _049///er: 1///pgr: NA///her2: NA///size: 2.2///node: 0///age: 60///grade: 1///t.dmfs: 3570///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6055///dataset: MAINZ///series: MAINZ///id: BC6055: Mainz_N0_untreated_frozen _050///er: 1///pgr: NA///her2: NA///size: 1.4///node: 0///age: 62///grade: 2///t.dmfs: 3930///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
  samplename: MAINZ_BC6056///dataset: MAINZ///series: MAINZ///id: BC6056wdh: Mainz_N0_untreated_frozen _051///er: 1///pgr: NA///her2: NA///size: 4///node: 0///age: 59///grade: 2///t.dmfs: 1980///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6057///dataset: MAINZ///series: MAINZ///id: BC6057: Mainz_N0_untreated_frozen _052///er: 1///pgr: NA///her2: NA///size: 2.2///node: 0///age: 78///grade: 2///t.dmfs: 2730///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6058///dataset: MAINZ///series: MAINZ///id: BC6058: Mainz_N0_untreated_frozen _053///er: 1///pgr: NA///her2: NA///size: 1.6///node: 0///age: 59///grade: 2///t.dmfs: 2940///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6059///dataset: MAINZ///series: MAINZ///id: BC6059: Mainz_N0_untreated_frozen _054///er: 1///pgr: NA///her2: NA///size: 2.5///node: 0///age: 42///grade: 2///t.dmfs: 1200///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
  samplename: MAINZ_BC6060///dataset: MAINZ///series: MAINZ///id: BC6060wdh: Mainz_N0_untreated_frozen _055///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 56///grade: 3///t.dmfs: 3600///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6061///dataset: MAINZ///series: MAINZ///id: BC6061: Mainz_N0_untreated_frozen _056///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 81///grade: 2///t.dmfs: 2370///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
samplename: MAINZ_BC6062///dataset: MAINZ///series: MAINZ///id: BC6062wdh: Mainz_N0_untreated_frozen _057///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 46///grade: 3///t.dmfs: 3660///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6063///dataset: MAINZ///series: MAINZ///id: BC6063: Mainz_N0_untreated_frozen _058///er: 1///pgr: NA///her2: NA///size: 4///node: 0///age: 64///grade: 2///t.dmfs: 3570///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
 samplename: MAINZ_BC6064///dataset: MAINZ///series: MAINZ///id: BC6064_2: Mainz_N0_untreated_frozen _059///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 64///grade: 2///t.dmfs: 1320///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6065///dataset: MAINZ///series: MAINZ///id: BC6065: Mainz_N0_untreated_frozen _060///er: 1///pgr: NA///her2: NA///size: 1.6///node: 0///age: 46///grade: 2///t.dmfs: 3390///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
samplename: MAINZ_BC6066///dataset: MAINZ///series: MAINZ///id: BC6066wdh: Mainz_N0_untreated_frozen _061///er: 1///pgr: NA///her2: NA///size: 2.2///node: 0///age: 67///grade: 2///t.dmfs: 3510///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6067///dataset: MAINZ///series: MAINZ///id: BC6067: Mainz_N0_untreated_frozen _062///er: 1///pgr: NA///her2: NA///size: 2.5///node: 0///age: 68///grade: 2///t.dmfs: 1980///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6068///dataset: MAINZ///series: MAINZ///id: BC6068: Mainz_N0_untreated_frozen _063///er: 0///pgr: NA///her2: NA///size: 4///node: 0///age: 40///grade: 3///t.dmfs: 3450///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6069///dataset: MAINZ///series: MAINZ///id: BC6069: Mainz_N0_untreated_frozen _064///er: 1///pgr: NA///her2: NA///size: 2.4///node: 0///age: 49///grade: 2///t.dmfs: 1920///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6070///dataset: MAINZ///series: MAINZ///id: BC6070: Mainz_N0_untreated_frozen _065///er: 1///pgr: NA///her2: NA///size: 1.4///node: 0///age: 71///grade: 2///t.dmfs: 3540///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6071///dataset: MAINZ///series: MAINZ///id: BC6071: Mainz_N0_untreated_frozen _066///er: 1///pgr: NA///her2: NA///size: 1.3///node: 0///age: 79///grade: 2///t.dmfs: 2790///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6072///dataset: MAINZ///series: MAINZ///id: BC6072: Mainz_N0_untreated_frozen _067///er: 1///pgr: NA///her2: NA///size: 2.7///node: 0///age: 58///grade: 2///t.dmfs: 2970///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6073///dataset: MAINZ///series: MAINZ///id: BC6073: Mainz_N0_untreated_frozen _068///er: 0///pgr: NA///her2: NA///size: 2.2///node: 0///age: 52///grade: 3///t.dmfs: 3180///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6074///dataset: MAINZ///series: MAINZ///id: BC6074: Mainz_N0_untreated_frozen _069///er: 1///pgr: NA///her2: NA///size: 1.3///node: 0///age: 64///grade: 2///t.dmfs: 2580///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6075///dataset: MAINZ///series: MAINZ///id: BC6075: Mainz_N0_untreated_frozen _070///er: 1///pgr: NA///her2: NA///size: 4///node: 0///age: 78///grade: 2///t.dmfs: 1530///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6076///dataset: MAINZ///series: MAINZ///id: BC6076: Mainz_N0_untreated_frozen _071///er: 1///pgr: NA///her2: NA///size: 2.4///node: 0///age: 79///grade: 1///t.dmfs: 3660///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6077///dataset: MAINZ///series: MAINZ///id: BC6077: Mainz_N0_untreated_frozen _072///er: 1///pgr: NA///her2: NA///size: 2.2///node: 0///age: 63///grade: 2///t.dmfs: 3840///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6078///dataset: MAINZ///series: MAINZ///id: BC6078: Mainz_N0_untreated_frozen _073///er: 1///pgr: NA///her2: NA///size: 1.1///node: 0///age: 54///grade: 1///t.dmfs: 3720///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6079///dataset: MAINZ///series: MAINZ///id: BC6079: Mainz_N0_untreated_frozen _074///er: 0///pgr: NA///her2: NA///size: 1.7///node: 0///age: 67///grade: 2///t.dmfs: 3420///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6080///dataset: MAINZ///series: MAINZ///id: BC6080: Mainz_N0_untreated_frozen _075///er: 1///pgr: NA///her2: NA///size: 1.5///node: 0///age: 47///grade: 2///t.dmfs: 3390///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6081///dataset: MAINZ///series: MAINZ///id: BC6081: Mainz_N0_untreated_frozen _076///er: 0///pgr: NA///her2: NA///size: 0.5///node: 0///age: 48///grade: 3///t.dmfs: 3660///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6082///dataset: MAINZ///series: MAINZ///id: BC6082: Mainz_N0_untreated_frozen _077///er: 1///pgr: NA///her2: NA///size: 0.9///node: 0///age: 66///grade: 2///t.dmfs: 3720///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6083///dataset: MAINZ///series: MAINZ///id: BC6083: Mainz_N0_untreated_frozen _078///er: 1///pgr: NA///her2: NA///size: 3.4///node: 0///age: 61///grade: 2///t.dmfs: 3450///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6084///dataset: MAINZ///series: MAINZ///id: BC6084: Mainz_N0_untreated_frozen _079///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 64///grade: 2///t.dmfs: 1350///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6085///dataset: MAINZ///series: MAINZ///id: BC6085: Mainz_N0_untreated_frozen _080///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 76///grade: 2///t.dmfs: 2100///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6086///dataset: MAINZ///series: MAINZ///id: BC6086: Mainz_N0_untreated_frozen _081///er: 1///pgr: NA///her2: NA///size: 0.1///node: 0///age: 49///grade: 2///t.dmfs: 4230///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
      samplename: MAINZ_BC6087///dataset: MAINZ///series: MAINZ///id: BC6087: Mainz_N0_untreated_frozen _082///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 43///grade: 3///t.dmfs: 240///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6088///dataset: MAINZ///series: MAINZ///id: BC6088: Mainz_N0_untreated_frozen _083///er: 1///pgr: NA///her2: NA///size: 0.2///node: 0///age: 69///grade: 1///t.dmfs: 4380///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6089///dataset: MAINZ///series: MAINZ///id: BC6089: Mainz_N0_untreated_frozen _084///er: 1///pgr: NA///her2: NA///size: 1.8///node: 0///age: 76///grade: 2///t.dmfs: 3930///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6091///dataset: MAINZ///series: MAINZ///id: BC6091: Mainz_N0_untreated_frozen _085///er: 0///pgr: NA///her2: NA///size: 1.8///node: 0///age: 54///grade: 3///t.dmfs: 4440///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6092///dataset: MAINZ///series: MAINZ///id: BC6092: Mainz_N0_untreated_frozen _086///er: 1///pgr: NA///her2: NA///size: 3///node: 0///age: 47///grade: 1///t.dmfs: 3990///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6093///dataset: MAINZ///series: MAINZ///id: BC6093: Mainz_N0_untreated_frozen _087///er: 1///pgr: NA///her2: NA///size: 3.5///node: 0///age: 46///grade: 1///t.dmfs: 4770///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6094///dataset: MAINZ///series: MAINZ///id: BC6094: Mainz_N0_untreated_frozen _088///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 56///grade: 2///t.dmfs: 4470///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
    samplename: MAINZ_BC6095///dataset: MAINZ///series: MAINZ///id: BC6095: Mainz_N0_untreated_frozen _089///er: 1///pgr: NA///her2: NA///size: 1.5///node: 0///age: 62///grade: 3///t.dmfs: 900///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6096///dataset: MAINZ///series: MAINZ///id: BC6096: Mainz_N0_untreated_frozen _090///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 41///grade: 2///t.dmfs: 5550///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6097///dataset: MAINZ///series: MAINZ///id: BC6097: Mainz_N0_untreated_frozen _091///er: 1///pgr: NA///her2: NA///size: 1.7///node: 0///age: 75///grade: 1///t.dmfs: 1470///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6098///dataset: MAINZ///series: MAINZ///id: BC6098: Mainz_N0_untreated_frozen _092///er: 0///pgr: NA///her2: NA///size: 5///node: 0///age: 49///grade: 3///t.dmfs: 5970///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6099///dataset: MAINZ///series: MAINZ///id: BC6099: Mainz_N0_untreated_frozen _093///er: 0///pgr: NA///her2: NA///size: 3.5///node: 0///age: 55///grade: 3///t.dmfs: 2700///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6100///dataset: MAINZ///series: MAINZ///id: BC6100: Mainz_N0_untreated_frozen _094///er: 1///pgr: NA///her2: NA///size: 1.5///node: 0///age: 53///grade: 1///t.dmfs: 5430///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
  samplename: MAINZ_BC6101///dataset: MAINZ///series: MAINZ///id: BC6101wdh: Mainz_N0_untreated_frozen _095///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 44///grade: 2///t.dmfs: 5700///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
   samplename: MAINZ_BC6102///dataset: MAINZ///series: MAINZ///id: BC6102: Mainz_N0_untreated_frozen _096///er: 1///pgr: NA///her2: NA///size: 1.5///node: 0///age: 51///grade: 2///t.dmfs: 3660///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
     samplename: MAINZ_BC6103///dataset: MAINZ///series: MAINZ///id: BC6103: Mainz_N0_untreated_frozen _097///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 49///grade: 2///t.dmfs: 4560///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
    samplename: MAINZ_BC6104///dataset: MAINZ///series: MAINZ///id: BC6104: Mainz_N0_untreated_frozen _098///er: 0///pgr: NA///her2: NA///size: 3.5///node: 0///age: 55///grade: 2///t.dmfs: 510///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
    samplename: MAINZ_BC6105///dataset: MAINZ///series: MAINZ///id: BC6105: Mainz_N0_untreated_frozen _099///er: 0///pgr: NA///her2: NA///size: 3.5///node: 0///age: 25///grade: 3///t.dmfs: 690///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                                                                    1 
                                                                                                                                                                                                                                                                              (Other) 
                                                                                                                                                                                                                                                                                  101 

}}
\source{ http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE11121 }
\keyword{datasets}
