\name{ TCGA_eset }
\alias{ TCGA_eset }
\docType{data}
\title{ TCGA_eset }
\usage{data( TCGA_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: http://www.ncbi.nlm.nih.gov/pubmed/?term=17428335; http://www.ncbi.nlm.nih.gov/pubmed/?term=14612510 
  Title:  
  URL: Not available 
  PMIDs: 17428335, 14612510  
  No abstract available.
  notes:
   summary:     
      A gene set was identified that correctly predicted outcomes more effectively than using histological markers.
   version:     
      2015-04-27 20:24:27
   mapping.method:     
      maxRowVariance
   mapping.group:     
      EntrezGene.ID

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: A1BG A2M ... ARHGAP11A.2 (19504 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 19504 features, 1073 samples
Platform type: 
Overall survival time-to-event summary (in years):
Call: survfit(formula = Surv(time, cens) ~ -1)

   970 observations deleted due to missingness 
records   n.max n.start  events  median 0.95LCL 0.95UCL 
 103.00  103.00  103.00  103.00    3.92    3.08    4.91 

--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
TCGA-3C-AAAU TCGA-3C-AALI TCGA-3C-AALJ TCGA-3C-AALK TCGA-4H-AAAK TCGA-5L-AAT0 TCGA-5L-AAT1 TCGA-5T-A9QA TCGA-A1-A0SB 
           1            1            1            1            1            1            1            1            1 
TCGA-A1-A0SD TCGA-A1-A0SE TCGA-A1-A0SF TCGA-A1-A0SG TCGA-A1-A0SH TCGA-A1-A0SI TCGA-A1-A0SJ TCGA-A1-A0SK TCGA-A1-A0SM 
           1            1            1            1            1            1            1            1            1 
TCGA-A1-A0SN TCGA-A1-A0SO TCGA-A1-A0SP TCGA-A1-A0SQ TCGA-A2-A04N TCGA-A2-A04P TCGA-A2-A04Q TCGA-A2-A04R TCGA-A2-A04T 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A04U TCGA-A2-A04V TCGA-A2-A04W TCGA-A2-A04X TCGA-A2-A04Y TCGA-A2-A0CK TCGA-A2-A0CL TCGA-A2-A0CM TCGA-A2-A0CO 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0CP TCGA-A2-A0CQ TCGA-A2-A0CR TCGA-A2-A0CS TCGA-A2-A0CT TCGA-A2-A0CU TCGA-A2-A0CV TCGA-A2-A0CW TCGA-A2-A0CX 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0CY TCGA-A2-A0CZ TCGA-A2-A0D0 TCGA-A2-A0D1 TCGA-A2-A0D2 TCGA-A2-A0D3 TCGA-A2-A0D4 TCGA-A2-A0EM TCGA-A2-A0EN 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0EO TCGA-A2-A0EP TCGA-A2-A0EQ TCGA-A2-A0ER TCGA-A2-A0ES TCGA-A2-A0ET TCGA-A2-A0EU TCGA-A2-A0EV TCGA-A2-A0EW 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0EX TCGA-A2-A0EY TCGA-A2-A0ST TCGA-A2-A0SU TCGA-A2-A0SV TCGA-A2-A0SW TCGA-A2-A0SX TCGA-A2-A0SY TCGA-A2-A0T0 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0T1 TCGA-A2-A0T2 TCGA-A2-A0T3 TCGA-A2-A0T4 TCGA-A2-A0T5 TCGA-A2-A0T6 TCGA-A2-A0T7 TCGA-A2-A0YC TCGA-A2-A0YD 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0YE TCGA-A2-A0YF TCGA-A2-A0YG TCGA-A2-A0YH TCGA-A2-A0YI TCGA-A2-A0YJ TCGA-A2-A0YK TCGA-A2-A0YL TCGA-A2-A0YM 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0YT TCGA-A2-A1FV TCGA-A2-A1FW TCGA-A2-A1FX TCGA-A2-A1FZ TCGA-A2-A1G0 TCGA-A2-A1G1 TCGA-A2-A1G4 TCGA-A2-A1G6 
           1            1            1            1            1            1            1            1            1 
     (Other) 
         974 

alt_sample_name: 
TCGA-3C-AAAU TCGA-3C-AALI TCGA-3C-AALJ TCGA-3C-AALK TCGA-4H-AAAK TCGA-5L-AAT0 TCGA-5L-AAT1 TCGA-5T-A9QA TCGA-A1-A0SB 
           1            1            1            1            1            1            1            1            1 
TCGA-A1-A0SD TCGA-A1-A0SE TCGA-A1-A0SF TCGA-A1-A0SG TCGA-A1-A0SH TCGA-A1-A0SI TCGA-A1-A0SJ TCGA-A1-A0SK TCGA-A1-A0SM 
           1            1            1            1            1            1            1            1            1 
TCGA-A1-A0SN TCGA-A1-A0SO TCGA-A1-A0SP TCGA-A1-A0SQ TCGA-A2-A04N TCGA-A2-A04P TCGA-A2-A04Q TCGA-A2-A04R TCGA-A2-A04T 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A04U TCGA-A2-A04V TCGA-A2-A04W TCGA-A2-A04X TCGA-A2-A04Y TCGA-A2-A0CK TCGA-A2-A0CL TCGA-A2-A0CM TCGA-A2-A0CO 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0CP TCGA-A2-A0CQ TCGA-A2-A0CR TCGA-A2-A0CS TCGA-A2-A0CT TCGA-A2-A0CU TCGA-A2-A0CV TCGA-A2-A0CW TCGA-A2-A0CX 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0CY TCGA-A2-A0CZ TCGA-A2-A0D0 TCGA-A2-A0D1 TCGA-A2-A0D2 TCGA-A2-A0D3 TCGA-A2-A0D4 TCGA-A2-A0EM TCGA-A2-A0EN 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0EO TCGA-A2-A0EP TCGA-A2-A0EQ TCGA-A2-A0ER TCGA-A2-A0ES TCGA-A2-A0ET TCGA-A2-A0EU TCGA-A2-A0EV TCGA-A2-A0EW 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0EX TCGA-A2-A0EY TCGA-A2-A0ST TCGA-A2-A0SU TCGA-A2-A0SV TCGA-A2-A0SW TCGA-A2-A0SX TCGA-A2-A0SY TCGA-A2-A0T0 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0T1 TCGA-A2-A0T2 TCGA-A2-A0T3 TCGA-A2-A0T4 TCGA-A2-A0T5 TCGA-A2-A0T6 TCGA-A2-A0T7 TCGA-A2-A0YC TCGA-A2-A0YD 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0YE TCGA-A2-A0YF TCGA-A2-A0YG TCGA-A2-A0YH TCGA-A2-A0YI TCGA-A2-A0YJ TCGA-A2-A0YK TCGA-A2-A0YL TCGA-A2-A0YM 
           1            1            1            1            1            1            1            1            1 
TCGA-A2-A0YT TCGA-A2-A1FV TCGA-A2-A1FW TCGA-A2-A1FX TCGA-A2-A1FZ TCGA-A2-A1G0 TCGA-A2-A1G1 TCGA-A2-A1G4 TCGA-A2-A1G6 
           1            1            1            1            1            1            1            1            1 
     (Other) 
         974 

unique_patient_ID: 
001cef41-ff86-4d3f-a140-a647ac4b10a1 0045349c-69d9-4306-a403-c9c1fa836644 00807dae-9f4a-4fd1-aac2-82eb11bf2afb 
                                   1                                    1                                    1 
00a2d166-78c9-4687-a195-3d6315c27574 00B11CA8-8540-4A3D-B602-EC754B00230B 011b9b2d-ebe5-42bf-9662-d922faccc7a1 
                                   1                                    1                                    1 
01263518-5f7c-49dc-8d7e-84b0c03a6a63 0130d616-885e-4a6c-9d03-2f17dd692a05 01674b2c-5cf2-478f-84a1-f69c39f47bd4 
                                   1                                    1                                    1 
016caf42-4e19-4444-ab5d-6cf1e76c4afa 01eef340-598c-4205-a990-cec190ac2ca5 01f50abf-fc15-446e-9d07-edeecc545a32 
                                   1                                    1                                    1 
029CE650-5E5A-4100-8596-CD94300E7EF5 02BBB632-0F7F-439D-B8F0-C86A06237424 02bed00f-bef7-4fb7-b243-540354990e45 
                                   1                                    1                                    1 
02bf5203-f9cd-4c5a-97b4-e5584dc22325 02e6f999-24f2-4ad2-9b92-df3ee2b082e5 02f5ae33-a563-4ecb-9e33-dfa500a44931 
                                   1                                    1                                    1 
03565036-b6d3-432f-8d9a-7099818e0788 035cb359-eac3-484a-9e88-3ddc739440a3 03ab1e9f-4918-404d-b178-3a4fd929c5e8 
                                   1                                    1                                    1 
03c143e0-d8a1-4d60-a4a3-df0501fc6b6e 03da7a65-fa2f-42cb-a709-5fe8736f21d3 045c13ef-3db7-4adf-b0a3-23338f0479f3 
                                   1                                    1                                    1 
048549CF-D0A5-4743-A0A1-2004F7CC2B08 0491063c-14a7-4104-8002-1459126332f3 04af44fc-c487-4005-b3cb-cb64815824a2 
                                   1                                    1                                    1 
05135396-df5c-49db-b34e-f4add5b65a2b 05506f4c-e701-4a9d-ae06-97f066aade43 058dcb78-bed1-40c3-a6c2-a0cc40065531 
                                   1                                    1                                    1 
05AF737D-89AA-49B6-97B2-28B8AC8BD2CF 0685edd2-ce1c-4e0e-8dda-393139af4223 06EF6E83-5583-45B9-A0BD-EDEA68943083 
                                   1                                    1                                    1 
0741b5db-4405-42ba-b63a-c6ee4f341480 074e8b06-3929-4896-9263-3b80ad6a355b 0772cdbe-8b0d-452b-8df1-bd70d1306363 
                                   1                                    1                                    1 
077506EA-8F11-405D-BB8E-3C9924FE54D0 0807435a-e75e-4e04-8e45-ed0cd49a841a 08306ede-e74b-4be9-b768-7fd98647b6ac 
                                   1                                    1                                    1 
084c6aec-94f7-4090-8f3f-59fa9e89721a 08740d7f-5a5e-4dfa-bd48-7fbf228a7a28 08da7c4c-3067-4bcf-9d7a-78566df72e69 
                                   1                                    1                                    1 
08de63a2-7b76-43c3-80dc-df748b1d81bc 0930E97A-72EA-4BD5-B2F0-50C76AF695F3 095c7985-3842-494f-b591-706ad1cd2133 
                                   1                                    1                                    1 
09765b0a-94f6-47d2-af56-93368084ac3a 09A5E9FD-D816-4F8C-BAA9-0E40BA607B16 09c2bc35-c21f-4aa4-ac30-0d8db02ad811 
                                   1                                    1                                    1 
09D6E904-7184-4788-8A19-04F8C46CB13F 0a017f15-1c6b-45e7-8d55-e0a71df1b2e8 0a229f95-0ea1-4d63-953e-74f5ec2f3028 
                                   1                                    1                                    1 
0a2a3529-f645-4967-9a58-89ee20b8bb62 0adf59c6-581a-475d-a2f4-40aa40060b5b 0B0F9938-CE1E-4814-8821-0132D137EB9E 
                                   1                                    1                                    1 
0b172e29-6166-4863-803d-a0a010887c11 0b23b5fe-6c7e-4297-bc55-f93fcb830ac4 0bc5744c-5fa3-45bb-87d0-70a02068b392 
                                   1                                    1                                    1 
0BF6E772-1AA1-4F54-9A50-C1414E2F22F3 0c1023c0-d2ce-4940-8047-7642f1dc3965 0c23c380-363c-474d-b64c-b47f612a8225 
                                   1                                    1                                    1 
0d03055c-186b-418a-b614-e2035757bc3b 0D146997-CFDC-4587-888C-E876FC455C80 0dc337fa-da8b-42c4-b9a7-fb76d81c161f 
                                   1                                    1                                    1 
0dca98b0-f43e-45b6-9a02-00092c78678c 0DD8DBC1-C48B-4E7C-B401-57101F724967 0e1d0d46-7a7b-4538-95dd-b9fc20a15aec 
                                   1                                    1                                    1 
0EC70E40-07DF-461D-8FF9-351240A0D454 0ed9ad26-7fdc-4d75-93f9-d029eee94993 0f64edec-0f1f-4025-8a53-75f9534f7828 
                                   1                                    1                                    1 
0f7ab21b-1fd0-45e0-9a47-17c4691eb84f 0fe1419e-a005-407c-8ae7-15c4c1579539 1026e0bf-5640-4fa5-b65f-4f11e8e76b18 
                                   1                                    1                                    1 
10C829EC-FD66-49C4-8AFE-AD3AE567372C 10c9be1e-2cc0-45e4-8d2c-c013ea63ef05 1111fb30-46ce-4722-98da-ec9ba2323d90 
                                   1                                    1                                    1 
1133B8A9-6B11-4511-B70A-F200E3B8B5DB 115bc72c-116f-40c1-be9b-e3dd7650efae 1174f6e4-ffbe-4e59-a000-8d861c968369 
                                   1                                    1                                    1 
124B693C-77DC-4FA8-B703-54E8B5054A92 125b0a9a-6cbe-4207-9178-e7ff03487e87 128d198e-9b22-427c-90db-3714455f3a17 
                                   1                                    1                                    1 
12d31787-e54d-4729-ace8-9141043c92b8 1347C23D-8D6E-4C15-9461-640249838C96 13b860c3-8048-4dc9-9eb0-7480bc660837 
                                   1                                    1                                    1 
13ca97a1-75bb-4a49-b8c3-181e5b2c9e86 1402082d-25c9-4148-8a1e-f58450b706bb 14267783-5624-4fe5-ba81-9d67f1017474 
                                   1                                    1                                    1 
14b95463-2108-4921-afc2-e29eef52b18f 14d1d84e-99e1-41ad-8892-3b4da3d4f948 1502c7d7-1535-4e56-9f34-30623acd50d5 
                                   1                                    1                                    1 
151f3677-6854-4265-9126-d92c646c67ae 1549dc64-3dab-43fc-96e9-b07d520957e1 1557183c-aedd-4df6-b98a-3654667f7b69 
                                   1                                    1                                    1 
157ec9e0-ca6e-4da1-9233-f976cf0db433 15FF4B02-BB69-4D24-8A07-B82F428E428D 161917b8-88ad-407b-ade6-d6b98478b359 
                                   1                                    1                                    1 
16368c32-2118-4fcf-8693-6c89995e49d8 16a8d916-ad46-4690-b1a5-3b2b323f4336 16FC3677-0393-4ED1-AD3F-C8355F056369 
                                   1                                    1                                    1 
                             (Other) 
                                 974 

sample_type: 
tumor 
 1073 

er: 
negative positive     NA's 
     233      790       50 

pgr: 
negative positive     NA's 
     334      686       53 

her2: 
negative positive     NA's 
     549      161      363 

age_at_initial_pathologic_diagnosis: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
  26.00   49.00   58.00   58.48   68.00   90.00 

days_to_death: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
  116.0   769.5  1430.0  1660.0  2446.0  4456.0     970 

vital_status: 
deceased   living 
     103      970 

batch: 
TCGA 
1073 

uncurated_author_metadata: 
                     bcr_patient_uuid: 001cef41-ff86-4d3f-a140-a647ac4b10a1///bcr_patient_barcode: TCGA-E2-A1IU///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 60///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0045349c-69d9-4306-a403-c9c1fa836644///bcr_patient_barcode: TCGA-A1-A0SB///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 70///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 00807dae-9f4a-4fd1-aac2-82eb11bf2afb///bcr_patient_barcode: TCGA-A2-A04W///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 50///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 00a2d166-78c9-4687-a195-3d6315c27574///bcr_patient_barcode: TCGA-AN-A0AM///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 56///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                 bcr_patient_uuid: 00B11CA8-8540-4A3D-B602-EC754B00230B///bcr_patient_barcode: TCGA-LL-A440///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: NO///age_at_diagnosis: 61///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 011b9b2d-ebe5-42bf-9662-d922faccc7a1///bcr_patient_barcode: TCGA-A7-A26E///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 71///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 01263518-5f7c-49dc-8d7e-84b0c03a6a63///bcr_patient_barcode: TCGA-A8-A07W///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 76///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0130d616-885e-4a6c-9d03-2f17dd692a05///bcr_patient_barcode: TCGA-D8-A1XY///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 74///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 01674b2c-5cf2-478f-84a1-f69c39f47bd4///bcr_patient_barcode: TCGA-EW-A1P7///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 59///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 016caf42-4e19-4444-ab5d-6cf1e76c4afa///bcr_patient_barcode: TCGA-AO-A128///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 61///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 01eef340-598c-4205-a990-cec190ac2ca5///bcr_patient_barcode: TCGA-B6-A0IQ///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 40///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 01f50abf-fc15-446e-9d07-edeecc545a32///bcr_patient_barcode: TCGA-E2-A155///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 58///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                bcr_patient_uuid: 029CE650-5E5A-4100-8596-CD94300E7EF5///bcr_patient_barcode: TCGA-E2-A574///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: YES///age_at_diagnosis: 44///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 02BBB632-0F7F-439D-B8F0-C86A06237424///bcr_patient_barcode: TCGA-AC-A2FB///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 65///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 02bed00f-bef7-4fb7-b243-540354990e45///bcr_patient_barcode: TCGA-A2-A0T1///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 55///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 02bf5203-f9cd-4c5a-97b4-e5584dc22325///bcr_patient_barcode: TCGA-AO-A0J8///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 61///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 02e6f999-24f2-4ad2-9b92-df3ee2b082e5///bcr_patient_barcode: TCGA-D8-A1JC///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 59///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 02f5ae33-a563-4ecb-9e33-dfa500a44931///bcr_patient_barcode: TCGA-A8-A093///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 61///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 03565036-b6d3-432f-8d9a-7099818e0788///bcr_patient_barcode: TCGA-AO-A0JC///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 64///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 035cb359-eac3-484a-9e88-3ddc739440a3///bcr_patient_barcode: TCGA-BH-A0DX///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 62///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 03ab1e9f-4918-404d-b178-3a4fd929c5e8///bcr_patient_barcode: TCGA-C8-A137///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 34///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
bcr_patient_uuid: 03c143e0-d8a1-4d60-a4a3-df0501fc6b6e///bcr_patient_barcode: TCGA-E9-A1N3///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 70///er_status_by_ihc: [Not Evaluated]///pr_status_by_ihc: [Not Evaluated]///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 03da7a65-fa2f-42cb-a709-5fe8736f21d3///bcr_patient_barcode: TCGA-D8-A146///history_neoadjuvant_treatment: Yes///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 57///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                            bcr_patient_uuid: 045c13ef-3db7-4adf-b0a3-23338f0479f3///bcr_patient_barcode: TCGA-B6-A0I6///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 991///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 49///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 048549CF-D0A5-4743-A0A1-2004F7CC2B08///bcr_patient_barcode: TCGA-AC-A3BB///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 46///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                            bcr_patient_uuid: 0491063c-14a7-4104-8002-1459126332f3///bcr_patient_barcode: TCGA-B6-A0IK///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 571///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 63///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
         bcr_patient_uuid: 04af44fc-c487-4005-b3cb-cb64815824a2///bcr_patient_barcode: TCGA-B6-A0IO///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 66///er_status_by_ihc: Positive///pr_status_by_ihc: Indeterminate///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
             bcr_patient_uuid: 05135396-df5c-49db-b34e-f4add5b65a2b///bcr_patient_barcode: TCGA-BH-A204///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 2534///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 80///er_status_by_ihc: [Not Evaluated]///pr_status_by_ihc: [Not Evaluated]///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 05506f4c-e701-4a9d-ae06-97f066aade43///bcr_patient_barcode: TCGA-AN-A0AT///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 62///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 058dcb78-bed1-40c3-a6c2-a0cc40065531///bcr_patient_barcode: TCGA-A2-A0T3///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 37///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 05AF737D-89AA-49B6-97B2-28B8AC8BD2CF///bcr_patient_barcode: TCGA-AC-A62Y///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Unknown]///age_at_diagnosis: 79///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                  bcr_patient_uuid: 0685edd2-ce1c-4e0e-8dda-393139af4223///bcr_patient_barcode: TCGA-AO-A03U///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 1793///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 31///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                bcr_patient_uuid: 06EF6E83-5583-45B9-A0BD-EDEA68943083///bcr_patient_barcode: TCGA-E2-A572///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: YES///age_at_diagnosis: 72///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                   bcr_patient_uuid: 0741b5db-4405-42ba-b63a-c6ee4f341480///bcr_patient_barcode: TCGA-BH-A1EY///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 538///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 79///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 074e8b06-3929-4896-9263-3b80ad6a355b///bcr_patient_barcode: TCGA-AR-A24S///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 61///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0772cdbe-8b0d-452b-8df1-bd70d1306363///bcr_patient_barcode: TCGA-BH-A0W4///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 46///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                 bcr_patient_uuid: 077506EA-8F11-405D-BB8E-3C9924FE54D0///bcr_patient_barcode: TCGA-AR-A5QN///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: YES///age_at_diagnosis: 68///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 0807435a-e75e-4e04-8e45-ed0cd49a841a///bcr_patient_barcode: TCGA-B6-A0IA///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 51///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 08306ede-e74b-4be9-b768-7fd98647b6ac///bcr_patient_barcode: TCGA-A2-A0T6///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 50///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 084c6aec-94f7-4090-8f3f-59fa9e89721a///bcr_patient_barcode: TCGA-A2-A0D1///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 76///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 08740d7f-5a5e-4dfa-bd48-7fbf228a7a28///bcr_patient_barcode: TCGA-B6-A0WZ///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 50///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 08da7c4c-3067-4bcf-9d7a-78566df72e69///bcr_patient_barcode: TCGA-B6-A0RE///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 61///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 08de63a2-7b76-43c3-80dc-df748b1d81bc///bcr_patient_barcode: TCGA-A1-A0SQ///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 45///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                 bcr_patient_uuid: 0930E97A-72EA-4BD5-B2F0-50C76AF695F3///bcr_patient_barcode: TCGA-GM-A3XN///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: YES///age_at_diagnosis: 44///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 095c7985-3842-494f-b591-706ad1cd2133///bcr_patient_barcode: TCGA-AO-A03M///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 29///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 09765b0a-94f6-47d2-af56-93368084ac3a///bcr_patient_barcode: TCGA-A7-A0CD///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 66///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                  bcr_patient_uuid: 09A5E9FD-D816-4F8C-BAA9-0E40BA607B16///bcr_patient_barcode: TCGA-LL-A7SZ///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: NO///age_at_diagnosis: 49///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 09c2bc35-c21f-4aa4-ac30-0d8db02ad811///bcr_patient_barcode: TCGA-A8-A09Q///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 83///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                  bcr_patient_uuid: 09D6E904-7184-4788-8A19-04F8C46CB13F///bcr_patient_barcode: TCGA-BH-AB28///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: NO///age_at_diagnosis: 53///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 0a017f15-1c6b-45e7-8d55-e0a71df1b2e8///bcr_patient_barcode: TCGA-A8-A08F///history_neoadjuvant_treatment: Yes///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 59///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0a229f95-0ea1-4d63-953e-74f5ec2f3028///bcr_patient_barcode: TCGA-AO-A0JD///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 59///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 0a2a3529-f645-4967-9a58-89ee20b8bb62///bcr_patient_barcode: TCGA-E2-A15E///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 40///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0adf59c6-581a-475d-a2f4-40aa40060b5b///bcr_patient_barcode: TCGA-A8-A09T///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 68///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                  bcr_patient_uuid: 0B0F9938-CE1E-4814-8821-0132D137EB9E///bcr_patient_barcode: TCGA-LD-A7W6///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: NO///age_at_diagnosis: 54///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 0b172e29-6166-4863-803d-a0a010887c11///bcr_patient_barcode: TCGA-B6-A0RO///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 71///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 0b23b5fe-6c7e-4297-bc55-f93fcb830ac4///bcr_patient_barcode: TCGA-D8-A27R///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 41///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0bc5744c-5fa3-45bb-87d0-70a02068b392///bcr_patient_barcode: TCGA-BH-A0DG///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 30///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 0BF6E772-1AA1-4F54-9A50-C1414E2F22F3///bcr_patient_barcode: TCGA-AR-A2LM///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 49///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 0c1023c0-d2ce-4940-8047-7642f1dc3965///bcr_patient_barcode: TCGA-BH-A0HB///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 55///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0c23c380-363c-474d-b64c-b47f612a8225///bcr_patient_barcode: TCGA-C8-A1HG///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 50///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0d03055c-186b-418a-b614-e2035757bc3b///bcr_patient_barcode: TCGA-D8-A1JS///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 77///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 0D146997-CFDC-4587-888C-E876FC455C80///bcr_patient_barcode: TCGA-A2-A4RW///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Unknown]///age_at_diagnosis: 49///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0dc337fa-da8b-42c4-b9a7-fb76d81c161f///bcr_patient_barcode: TCGA-A1-A0SN///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 50///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0dca98b0-f43e-45b6-9a02-00092c78678c///bcr_patient_barcode: TCGA-D8-A27H///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 72///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                bcr_patient_uuid: 0DD8DBC1-C48B-4E7C-B401-57101F724967///bcr_patient_barcode: TCGA-S3-AA11///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: YES///age_at_diagnosis: 67///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 0e1d0d46-7a7b-4538-95dd-b9fc20a15aec///bcr_patient_barcode: TCGA-E2-A1BD///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 53///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                 bcr_patient_uuid: 0EC70E40-07DF-461D-8FF9-351240A0D454///bcr_patient_barcode: TCGA-A7-A5ZV///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: NO///age_at_diagnosis: 62///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0ed9ad26-7fdc-4d75-93f9-d029eee94993///bcr_patient_barcode: TCGA-A1-A0SG///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 61///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 0f64edec-0f1f-4025-8a53-75f9534f7828///bcr_patient_barcode: TCGA-BH-A0H9///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 69///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 0f7ab21b-1fd0-45e0-9a47-17c4691eb84f///bcr_patient_barcode: TCGA-E2-A15S///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 34///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 0fe1419e-a005-407c-8ae7-15c4c1579539///bcr_patient_barcode: TCGA-E2-A1II///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 51///er_status_by_ihc: Negative///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 1026e0bf-5640-4fa5-b65f-4f11e8e76b18///bcr_patient_barcode: TCGA-D8-A1XS///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 48///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                  bcr_patient_uuid: 10C829EC-FD66-49C4-8AFE-AD3AE567372C///bcr_patient_barcode: TCGA-BH-A42T///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: NO///age_at_diagnosis: 75///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 10c9be1e-2cc0-45e4-8d2c-c013ea63ef05///bcr_patient_barcode: TCGA-AN-A0FS///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 55///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                  bcr_patient_uuid: 1111fb30-46ce-4722-98da-ec9ba2323d90///bcr_patient_barcode: TCGA-AO-A03O///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 2483///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 69///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 1133B8A9-6B11-4511-B70A-F200E3B8B5DB///bcr_patient_barcode: TCGA-GM-A2D9///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 1812///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 69///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 115bc72c-116f-40c1-be9b-e3dd7650efae///bcr_patient_barcode: TCGA-D8-A1JE///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 62///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 1174f6e4-ffbe-4e59-a000-8d861c968369///bcr_patient_barcode: TCGA-AO-A03L///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 34///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 124B693C-77DC-4FA8-B703-54E8B5054A92///bcr_patient_barcode: TCGA-HN-A2NL///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Unknown]///age_at_diagnosis: 56///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 125b0a9a-6cbe-4207-9178-e7ff03487e87///bcr_patient_barcode: TCGA-AR-A24O///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 43///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                   bcr_patient_uuid: 128d198e-9b22-427c-90db-3714455f3a17///bcr_patient_barcode: TCGA-A2-A0T2///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 255///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 66///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 12d31787-e54d-4729-ace8-9141043c92b8///bcr_patient_barcode: TCGA-A2-A0ST///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 62///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                 bcr_patient_uuid: 1347C23D-8D6E-4C15-9461-640249838C96///bcr_patient_barcode: TCGA-LL-A6FR///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: NO///age_at_diagnosis: 50///er_status_by_ihc: Negative///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 13b860c3-8048-4dc9-9eb0-7480bc660837///bcr_patient_barcode: TCGA-AN-A0FX///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 52///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 13ca97a1-75bb-4a49-b8c3-181e5b2c9e86///bcr_patient_barcode: TCGA-AR-A0TT///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 53///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
              bcr_patient_uuid: 1402082d-25c9-4148-8a1e-f58450b706bb///bcr_patient_barcode: TCGA-B6-A0IP///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 74///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 14267783-5624-4fe5-ba81-9d67f1017474///bcr_patient_barcode: TCGA-BH-A0DP///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 60///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 14b95463-2108-4921-afc2-e29eef52b18f///bcr_patient_barcode: TCGA-A2-A0YT///history_neoadjuvant_treatment: Yes///vital_status: Dead///death_days_to: 723///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 56///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 14d1d84e-99e1-41ad-8892-3b4da3d4f948///bcr_patient_barcode: TCGA-BH-A1FL///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 1673///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 69///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 1502c7d7-1535-4e56-9f34-30623acd50d5///bcr_patient_barcode: TCGA-B6-A1KC///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 67///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 151f3677-6854-4265-9126-d92c646c67ae///bcr_patient_barcode: TCGA-B6-A0RM///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 2373///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 57///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 1549dc64-3dab-43fc-96e9-b07d520957e1///bcr_patient_barcode: TCGA-D8-A27F///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 40///er_status_by_ihc: Negative///pr_status_by_ihc: Negative///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 1557183c-aedd-4df6-b98a-3654667f7b69///bcr_patient_barcode: TCGA-A8-A09I///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 84///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 157ec9e0-ca6e-4da1-9233-f976cf0db433///bcr_patient_barcode: TCGA-B6-A0IB///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 3941///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 64///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                    bcr_patient_uuid: 15FF4B02-BB69-4D24-8A07-B82F428E428D///bcr_patient_barcode: TCGA-A7-A3IZ///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 62///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 161917b8-88ad-407b-ade6-d6b98478b359///bcr_patient_barcode: TCGA-E9-A22H///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 42///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Positive///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                     bcr_patient_uuid: 16368c32-2118-4fcf-8693-6c89995e49d8///bcr_patient_barcode: TCGA-A1-A0SI///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 52///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Negative///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                           bcr_patient_uuid: 16a8d916-ad46-4690-b1a5-3b2b323f4336///bcr_patient_barcode: TCGA-B6-A0IN///history_neoadjuvant_treatment: No///vital_status: Dead///death_days_to: 2573///radiation_treatment_adjuvant: [Not Available]///age_at_diagnosis: 45///er_status_by_ihc: Positive///pr_status_by_ihc: Negative///her2_status_by_ihc: [Not Evaluated]///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                 bcr_patient_uuid: 16FC3677-0393-4ED1-AD3F-C8355F056369///bcr_patient_barcode: TCGA-5L-AAT1///history_neoadjuvant_treatment: No///vital_status: Alive///death_days_to: [Not Applicable]///radiation_treatment_adjuvant: NO///age_at_diagnosis: 63///er_status_by_ihc: Positive///pr_status_by_ihc: Positive///her2_status_by_ihc: Equivocal///clinical_N: [Not Applicable]///tumor_tissue_site: Breast 
                                                                                                                                                                                                                                                                                                                                                                                                                                     1 
                                                                                                                                                                                                                                                                                                                                                                                                                               (Other) 
                                                                                                                                                                                                                                                                                                                                                                                                                                   974 

}}
\source{ Not available }
\keyword{datasets}
