\name{ VDX_eset }
\alias{ VDX_eset }
\docType{data}
\title{ VDX_eset }
\usage{data( VDX_eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name:  
  Laboratory:  
  Contact information: NA 
  Title:  
  URL: NA 
  PMIDs: NA 

featureData(eset):
An object of class 'AnnotatedDataFrame'
  featureNames: 1007_s_at 1053_at ... AFFX-HUMISGF3A/M97935_MB_at (21169 total)
  varLabels: probeset gene EntrezGene.ID best_probe
  varMetadata: labelDescription

}}
\details{
\preformatted{
assayData: 21169 features, 344 samples
Platform type: 
--------------------------- 
Available sample meta-data: 
--------------------------- 

sample_name: 
 VDX_100  VDX_103  VDX_104  VDX_105  VDX_106 VDX_1060 VDX_1069  VDX_107  VDX_108  VDX_109   VDX_11  VDX_110  VDX_111  VDX_112 
       1        1        1        1        1        1        1        1        1        1        1        1        1        1 
 VDX_113  VDX_114  VDX_115  VDX_118  VDX_120  VDX_121  VDX_122 VDX_1222  VDX_123  VDX_124  VDX_125  VDX_126  VDX_130  VDX_131 
       1        1        1        1        1        1        1        1        1        1        1        1        1        1 
 VDX_133  VDX_134  VDX_135  VDX_136  VDX_137  VDX_138  VDX_139   VDX_14  VDX_140  VDX_141 VDX_1412 VDX_1416 VDX_1455 VDX_1482 
       1        1        1        1        1        1        1        1        1        1        1        1        1        1 
VDX_1487 VDX_1497   VDX_15 VDX_1501 VDX_1524 VDX_1535 VDX_1555 VDX_1563 VDX_1568 VDX_1571 VDX_1582 VDX_1589 VDX_1593 VDX_1622 
       1        1        1        1        1        1        1        1        1        1        1        1        1        1 
VDX_1650 VDX_1661   VDX_17   VDX_18 VDX_1821 VDX_1828 VDX_1853 VDX_1858 VDX_1871   VDX_19 VDX_1938 VDX_1939 VDX_1949 VDX_1966 
       1        1        1        1        1        1        1        1        1        1        1        1        1        1 
VDX_1967 VDX_1976 VDX_1978   VDX_20  VDX_200 VDX_2001 VDX_2007  VDX_201 VDX_2021 VDX_2024 VDX_2025  VDX_203 VDX_2038  VDX_204 
       1        1        1        1        1        1        1        1        1        1        1        1        1        1 
 VDX_205   VDX_21  VDX_212  VDX_213  VDX_215  VDX_216  VDX_217   VDX_22  VDX_222  VDX_223  VDX_226  VDX_230  VDX_231  VDX_233 
       1        1        1        1        1        1        1        1        1        1        1        1        1        1 
 VDX_234  (Other) 
       1      245 

alt_sample_name: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
    3.0   122.8   605.5   575.7   836.5  2038.0 

sample_type: 
tumor 
  344 

er: 
negative positive 
     135      209 

tumor_size: 
   1    2    3    4 NA's 
 146  132    5    3   58 

N: 
  0 
344 

age_at_initial_pathologic_diagnosis: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
  26.00   44.00   52.00   53.88   63.00   83.00      58 

grade: 
   1    2    3 NA's 
   7   42  148  147 

dmfs_days: 
   Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
     61    1254    2616    2377    3285    5201 

dmfs_status: 
deceased_or_recurrence    living_norecurrence 
                   118                    226 

treatment: 
untreated 
      344 

batch: 
 VDX VDXN 
 286   58 

uncurated_author_metadata: 
     samplename: VDX_100///dataset: VDX///series: VDX///id: 100///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 48///grade: 2///t.dmfs: 1186///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_103///dataset: VDX///series: VDX///id: 103///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 41///grade: 3///t.dmfs: 2950///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_104///dataset: VDX///series: VDX///id: 104///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 53///grade: NA///t.dmfs: 335///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_105///dataset: VDX///series: VDX///id: 105///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 50///grade: 3///t.dmfs: 3011///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_106///dataset: VDX///series: VDX///id: 106///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 46///grade: NA///t.dmfs: 1217///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1060///dataset: VDX///series: VDXN///id: 1060///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2168///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
 samplename: VDX_1069///dataset: VDX///series: VDXN///id: 1069///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 719///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_107///dataset: VDX///series: VDX///id: 107///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 53///grade: 3///t.dmfs: 2859///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_108///dataset: VDX///series: VDX///id: 108///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 66///grade: NA///t.dmfs: 852///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_109///dataset: VDX///series: VDX///id: 109///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 65///grade: 3///t.dmfs: 2646///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
       samplename: VDX_11///dataset: VDX///series: VDX///id: 11///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 69///grade: 3///t.dmfs: 3315///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_110///dataset: VDX///series: VDX///id: 110///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 73///grade: 3///t.dmfs: 243///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_111///dataset: VDX///series: VDX///id: 111///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 60///grade: NA///t.dmfs: 2981///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_112///dataset: VDX///series: VDX///id: 112///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 47///grade: NA///t.dmfs: 517///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_113///dataset: VDX///series: VDX///id: 113///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 47///grade: 3///t.dmfs: 3072///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_114///dataset: VDX///series: VDX///id: 114///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 62///grade: 2///t.dmfs: 2616///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_115///dataset: VDX///series: VDX///id: 115///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 61///grade: 3///t.dmfs: 456///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_118///dataset: VDX///series: VDX///id: 118///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 78///grade: NA///t.dmfs: 2890///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
  samplename: VDX_120///dataset: VDX///series: VDXN///id: 120///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2919///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_121///dataset: VDX///series: VDX///id: 121///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 46///grade: 3///t.dmfs: 700///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_122///dataset: VDX///series: VDX///id: 122///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 54///grade: 2///t.dmfs: 3163///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
 samplename: VDX_1222///dataset: VDX///series: VDXN///id: 1222///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 218///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
  samplename: VDX_123///dataset: VDX///series: VDXN///id: 123///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2687///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_124///dataset: VDX///series: VDX///id: 124///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 50///grade: 3///t.dmfs: 243///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_125///dataset: VDX///series: VDX///id: 125///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 54///grade: NA///t.dmfs: 2829///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_126///dataset: VDX///series: VDX///id: 126///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 49///grade: NA///t.dmfs: 1125///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_130///dataset: VDX///series: VDX///id: 130///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 55///grade: 3///t.dmfs: 791///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
  samplename: VDX_131///dataset: VDX///series: VDXN///id: 131///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2960///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_133///dataset: VDX///series: VDX///id: 133///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 72///grade: NA///t.dmfs: 2646///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_134///dataset: VDX///series: VDX///id: 134///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 57///grade: NA///t.dmfs: 852///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_135///dataset: VDX///series: VDX///id: 135///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 28///grade: NA///t.dmfs: 2555///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_136///dataset: VDX///series: VDX///id: 136///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 69///grade: 3///t.dmfs: 760///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_137///dataset: VDX///series: VDX///id: 137///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 69///grade: 3///t.dmfs: 973///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_138///dataset: VDX///series: VDX///id: 138///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 59///grade: NA///t.dmfs: 1430///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_139///dataset: VDX///series: VDX///id: 139///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 47///grade: 3///t.dmfs: 3376///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
        samplename: VDX_14///dataset: VDX///series: VDX///id: 14///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 68///grade: 3///t.dmfs: 426///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_140///dataset: VDX///series: VDX///id: 140///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 37///grade: NA///t.dmfs: 578///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_141///dataset: VDX///series: VDX///id: 141///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 74///grade: 3///t.dmfs: 760///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1412///dataset: VDX///series: VDXN///id: 1412///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3885///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1416///dataset: VDX///series: VDXN///id: 1416///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3237///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1455///dataset: VDX///series: VDXN///id: 1455///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 1956///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1482///dataset: VDX///series: VDXN///id: 1482///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2626///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1487///dataset: VDX///series: VDXN///id: 1487///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2146///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1497///dataset: VDX///series: VDXN///id: 1497///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3633///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
       samplename: VDX_15///dataset: VDX///series: VDX///id: 15///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 44///grade: 3///t.dmfs: 3011///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1501///dataset: VDX///series: VDXN///id: 1501///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2242///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1524///dataset: VDX///series: VDXN///id: 1524///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2079///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1535///dataset: VDX///series: VDXN///id: 1535///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 1267///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
 samplename: VDX_1555///dataset: VDX///series: VDXN///id: 1555///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 279///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1563///dataset: VDX///series: VDXN///id: 1563///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3141///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1568///dataset: VDX///series: VDXN///id: 1568///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3499///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1571///dataset: VDX///series: VDXN///id: 1571///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2322///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1582///dataset: VDX///series: VDXN///id: 1582///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2927///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1589///dataset: VDX///series: VDXN///id: 1589///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3957///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1593///dataset: VDX///series: VDXN///id: 1593///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3858///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1622///dataset: VDX///series: VDXN///id: 1622///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3989///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1650///dataset: VDX///series: VDXN///id: 1650///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2675///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1661///dataset: VDX///series: VDXN///id: 1661///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2459///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
       samplename: VDX_17///dataset: VDX///series: VDX///id: 17///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 51///grade: 3///t.dmfs: 4167///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
       samplename: VDX_18///dataset: VDX///series: VDX///id: 18///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 32///grade: 3///t.dmfs: 1034///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1821///dataset: VDX///series: VDXN///id: 1821///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2672///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1828///dataset: VDX///series: VDXN///id: 1828///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3126///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1853///dataset: VDX///series: VDXN///id: 1853///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 1905///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1858///dataset: VDX///series: VDXN///id: 1858///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2696///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1871///dataset: VDX///series: VDXN///id: 1871///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2620///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
       samplename: VDX_19///dataset: VDX///series: VDX///id: 19///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 66///grade: NA///t.dmfs: 973///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1938///dataset: VDX///series: VDXN///id: 1938///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2734///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1939///dataset: VDX///series: VDXN///id: 1939///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 4608///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1949///dataset: VDX///series: VDXN///id: 1949///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3390///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1966///dataset: VDX///series: VDXN///id: 1966///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 1870///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1967///dataset: VDX///series: VDXN///id: 1967///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2425///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1976///dataset: VDX///series: VDXN///id: 1976///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2212///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_1978///dataset: VDX///series: VDXN///id: 1978///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2098///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
       samplename: VDX_20///dataset: VDX///series: VDX///id: 20///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 34///grade: 3///t.dmfs: 3893///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_200///dataset: VDX///series: VDX///id: 200///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 68///grade: 3///t.dmfs: 3285///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
 samplename: VDX_2001///dataset: VDX///series: VDXN///id: 2001///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 133///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_2007///dataset: VDX///series: VDXN///id: 2007///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2774///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_201///dataset: VDX///series: VDX///id: 201///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 70///grade: 2///t.dmfs: 3437///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_2021///dataset: VDX///series: VDXN///id: 2021///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2772///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
 samplename: VDX_2024///dataset: VDX///series: VDXN///id: 2024///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 657///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_2025///dataset: VDX///series: VDXN///id: 2025///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 2522///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_203///dataset: VDX///series: VDX///id: 203///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 43///grade: NA///t.dmfs: 882///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
samplename: VDX_2038///dataset: VDX///series: VDXN///id: 2038///er: 0///pgr: NA///her2: NA///size: NA///node: 0///age: NA///grade: NA///t.dmfs: 3107///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_204///dataset: VDX///series: VDX///id: 204///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 66///grade: NA///t.dmfs: 730///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_205///dataset: VDX///series: VDX///id: 205///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 47///grade: 3///t.dmfs: 700///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
       samplename: VDX_21///dataset: VDX///series: VDX///id: 21///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 46///grade: NA///t.dmfs: 426///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_212///dataset: VDX///series: VDX///id: 212///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 46///grade: 3///t.dmfs: 213///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_213///dataset: VDX///series: VDX///id: 213///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 32///grade: 3///t.dmfs: 487///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_215///dataset: VDX///series: VDX///id: 215///er: 0///pgr: NA///her2: NA///size: 2///node: 0///age: 53///grade: 3///t.dmfs: 183///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_216///dataset: VDX///series: VDX///id: 216///er: 0///pgr: NA///her2: NA///size: 1///node: 0///age: 46///grade: 3///t.dmfs: 395///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_217///dataset: VDX///series: VDX///id: 217///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 50///grade: 3///t.dmfs: 578///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
       samplename: VDX_22///dataset: VDX///series: VDX///id: 22///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 32///grade: 3///t.dmfs: 3954///e.dmfs: 0///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_222///dataset: VDX///series: VDX///id: 222///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 70///grade: 3///t.dmfs: 1125///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_223///dataset: VDX///series: VDX///id: 223///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 53///grade: NA///t.dmfs: 700///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_226///dataset: VDX///series: VDX///id: 226///er: 1///pgr: NA///her2: NA///size: 2///node: 0///age: 59///grade: 3///t.dmfs: 1156///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
      samplename: VDX_230///dataset: VDX///series: VDX///id: 230///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 26///grade: 2///t.dmfs: 730///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_231///dataset: VDX///series: VDX///id: 231///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 49///grade: 3///t.dmfs: 1338///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
     samplename: VDX_233///dataset: VDX///series: VDX///id: 233///er: 1///pgr: NA///her2: NA///size: 3///node: 0///age: 37///grade: NA///t.dmfs: 152///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
    samplename: VDX_234///dataset: VDX///series: VDX///id: 234///er: 1///pgr: NA///her2: NA///size: 1///node: 0///age: 43///grade: NA///t.dmfs: 1125///e.dmfs: 1///t.rfs: NA///e.rfs: NA///t.os: NA///e.os: NA///tissue: 1///treatment: 0 
                                                                                                                                                                                                                                        1 
                                                                                                                                                                                                                                  (Other) 
                                                                                                                                                                                                                                      245 

}}
\keyword{datasets}
